### 2022-05-26 #4424 @editorialbot
[PRE REVIEW]: vSmartMOM.jl: an Open-Source Julia Package for Atmospheric Radiative Transfer and Remote Sensing Tools
**Submitting author:** @RupeshJey (<a href="http://orcid.org/0000-0003-0142-7367">Rupesh Jeyaram</a>)
**Repository:** https://github.com/RemoteSensingTools/vSmartMOM.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
