+++ test/rami/rami.jl
@@ -1,0 +1,164 @@
+using JSON
+using vSmartMOM
+using DelimitedFiles
+using Distributions
+
+const sentinel_band_to_wn = Dict([("2" , 1e7 ./ [492.4 492.5]),    # Blue
+                                  ("3" , 1e7 ./ [559.8 559.9]),    # Green
+                                  ("4" , 1e7 ./ [664.6 664.7]),    # Red
+                                  ("8A", 1e7 ./ [864.7 864.8]),    # Narrow NIR 
+                                  ("11", 1e7 ./ [1613.7 1613.8]),  # SWIR
+                                  ("12", 1e7 ./ [2202.4 2202.5])]) # SWIR 
+
+
+## 
+
+# Read in the standard atmosphere
+
+# atmosphere ascii file location 
+rami_atmos_fn = "/home/rjeyaram/vSmartMOM/test/rami/RAMI4ATM_AFGLUSstandard_ap_v1.0.txt"
+US_std_atmos = readdlm(rami_atmos_fn, ' ')
+
+## 
+
+# Starts with HOM00_LAM, HOM00_BLA, or HOM00_WHI for lambertian surfaces
+# Starts with HOM00_RPV, HOM00_RLI for Anisotropic surfaces
+# Starts with HOM45_LAM, HOM25_LAM, or HOM35_LAM for homogeneous discrete with isotropic background
+
+function produce_rami_results(experiment_name::String;
+                              rami_json::String = "/home/rjeyaram/vSmartMOM/test/rami/RAMI4ATM_experiments_v1.0.json",
+                              default_params::String = "/home/rjeyaram/vSmartMOM/test/test_parameters/PureRayleighParameters.yaml")
+
+    # Get rami scenario from experiment name
+    all_scenarios = JSON.parsefile(rami_json)
+    filter!(x->x["name"] == experiment_name, all_scenarios)
+    @assert length(all_scenarios) ≤ 1 "Multiple matching experiment names!"
+    @assert length(all_scenarios) ≥ 1 "Experiment name not found in JSON input"
+    curr_scenario = all_scenarios[1]["observations"][1]
+
+    # Get a default set of Rayleigh params
+    params = vSmartMOM.parameters_from_yaml(default_params)
+
+    # ########################################
+    # Change params fields to match rami input
+    # ########################################
+
+    # There are 7 rami categories to be mindful of: 
+    # atmosphere, name, canopy, illumination, measures, time, surface
+
+    rami_atmosphere = curr_scenario["atmosphere"]
+    rami_name = curr_scenario["name"]
+    rami_canopy = curr_scenario["canopy"]
+    rami_illumination = curr_scenario["illumination"]
+    rami_measures = curr_scenario["measures"]
+    rami_time = curr_scenario["time"]
+    rami_surface = curr_scenario["surface"]
+
+    # ########################################
+    # 1. Atmosphere
+    # ########################################
+
+    # aerosols 
+    if length(rami_atmosphere["aerosols"]) > 0
+
+        if startswith(rami_atmosphere["aerosols"][1]["name"], "D")
+
+            μ  = 0.0478666
+            σ  = 1.87411
+            # These refractive indices depend on the band. 
+            # See https://rami-benchmark.jrc.ec.europa.eu/_www/RAMI4ATM/down/RAMI4ATM_aerosols_v1.0/refractive_index/continental.txt
+            # Need to generalize
+            nᵣ = 1.4434925925925925
+            nᵢ = 0.0015797
+
+        elseif startswith(rami_atmosphere["aerosols"][1]["name"], "C")
+
+            μ  = 0.0807989
+            σ  = 1.50180
[…]
