+++ test/rami/RamiGas.yaml
@@ -1,0 +1,126 @@
+# =================================================================
+# Top-Level Radiative Transfer Parameters
+# =================================================================
+radiative_transfer: 
+  # Spectral bands (list of ν_start:ν_step:ν_end, in cm⁻¹)
+  spec_bands: 
+    - (1e7/773):0.02:(1e7/757)
+    - (1e7/1622):0.02:(1e7/1589)
+  #  - (1e7/2084):0.02:(1e7/2042)
+  # Bidirectional Reflectance Distribution Function (BRDF) per band
+  surface: 
+    - LambertianSurfaceScalar(0.05)
+    - LambertianSurfaceScalar(0.05)
+  #  - LambertianSurfaceScalar(0.05)
+  # Quadrature-scheme for stream calculations
+  quadrature_type:    GaussQuadHemisphere()
+  # Polarization type (Stokes_I, Stokes_IQ, Stokes_IQU, Stokes_IQUV)
+  polarization_type:  Stokes_IQU()
+  # Maximum m (number of Fourier moments to loop over)
+  max_m:              3
+  # Exclusion angle for forward peak in fitting procedure (degrees)
+  Δ_angle:            2.0
+  # Trunction length for legendre terms
+  l_trunc:            10
+  # Depolarization factor
+  depol:              0.03
+  # Floating point type for calculations (Float32, Float64)
+  float_type:         Float64
+  # Architecture (default_architecture, GPU(), CPU())
+  architecture:       default_architecture #default_architecture
+  
+# =================================================================
+# Simulation Geometry
+# =================================================================
+geometry: 
+  # Solar zenith angle (degrees)
+  sza:                32.4436
+  # Viewing zenith angles (degrees)
+  vza:                [0.072]
+  # Viewing azimuth angles (degrees)
+  vaz:                [0.0]
+  # Observation altitude (Pa)
+  obs_alt:            1000.0
+
+# =================================================================
+# Atmospheric Profile
+# =================================================================
+atmospheric_profile: 
+  # Temperature Profile (TOA to BOA, in K)
+  # Full-levels, specified at layer centers
+  T: [330.0, 270.0, 224.4, 201.95, 191.75, 187.65, 187.9, 193.75, 
+      203.5, 214.0, 226.45, 240.15, 253.9, 265.75, 270.65, 267.4, 
+      260.75, 253.85000000000002, 246.65, 239.7, 233.25, 228.25, 
+      225.25, 222.8, 221.1, 220.1, 219.1, 218.1, 217.14999999999998, 
+      216.7, 216.7, 216.7, 216.7, 216.7, 216.7, 216.7, 216.7, 216.75, 
+      220.05, 226.5, 232.95, 239.45, 245.95, 252.45, 258.95, 265.45, 271.95, 278.45, 284.95]
+  # Pressure Profile (TOA to BOA at each boundary, in hPa)
[…]
