### 2022-07-11 c0f6f4f
Change aerosol height profile to using Distributions.jl
