+++ test/rami/RamiNoGas.yaml
@@ -44,19 +44,39 @@ geometry:
 atmospheric_profile: 
   # Temperature Profile (TOA to BOA, in K)
   # Full-levels, specified at layer centers
-  T: [244.33, 251.34, 258.09, 264.25, 269.15, 
-      272.59, 274.07, 273.30, 269.65, 264.27, 258.11, 
-      251.52, 245.22, 239.20, 234.05, 229.71, 225.70, 
-      222.70, 220.62, 219.32, 217.93, 216.98, 217.10, 
-      218.35, 223.33, 234.19, 249.34, 264.12, 277.20, 
-      280.77, 282.60, 284.40, 285.80]
+  # Full-levels, specified at layer centers
+  T: [330.0, 270.0, 224.4, 201.95, 191.75, 187.65, 187.9, 193.75, 
+      203.5, 214.0, 226.45, 240.15, 253.9, 265.75, 270.65, 267.4, 
+      260.75, 253.85000000000002, 246.65, 239.7, 233.25, 228.25, 
+      225.25, 222.8, 221.1, 220.1, 219.1, 218.1, 217.14999999999998, 
+      216.7, 216.7, 216.7, 216.7, 216.7, 216.7, 216.7, 216.7, 216.75, 
+      220.05, 226.5, 232.95, 239.45, 245.95, 252.45, 258.95, 265.45, 271.95, 278.45, 284.95]
   # Pressure Profile (TOA to BOA at each boundary, in hPa)
   # Half-levels, specified at layer boundaries
-  p: [  0.14,   0.22,   0.30,   0.39,   0.53,   0.71,   
-        0.96,   1.28,   1.70,   2.27,   3.03,   4.03,  
-        5.44,   7.26,   9.67,  12.90,  17.23,  23.30,  
-       31.00,  42.07,  56.09,  74.78,  99.69, 131.00,
-      176.85, 236.64, 314.58, 418.87, 557.76, 735.00, 
-      800.12, 849.00, 912.00, 980.00]
+  p: [2.54e-5, 4.01e-5, 7.1e-5, 0.000145, 0.00032, 0.00076, 0.00184, 
+      0.00446, 0.0105, 0.024, 0.0522, 0.109, 0.219, 0.425, 0.7978, 
+      1.09, 1.491, 2.06, 2.871, 4.15, 5.746, 8.01, 11.97, 17.43, 
+      25.49, 29.72, 34.67, 40.47, 47.29, 55.29, 64.67, 75.65, 88.5, 
+      103.5, 121.1, 141.7, 165.8, 194.0, 227.0, 265.0, 308.0, 356.5, 
+      411.1, 472.2, 540.5, 616.6, 701.2, 795.0, 898.8, 1013.0]
   # Reduce profile to n layers
   profile_reduction: -1
+
+scattering:
+  # List of scattering aerosols and their properties
+  aerosols:       
+    - τ_ref:          0.2         # Reference τ
+      μ:              1.0           # effective radius (µm)
+      σ:              1.5           # geometric standard deviation (µm)
+      nᵣ:             1.3           # Real part of refractive index
+      nᵢ:             0.00000001    # Imag part of refractive index
+      p₀:             700.0         # Pressure peak (hPa)
+      σp:             50.0          # Pressure peak width (hPa)
+  # Maximum aerosol particle radius for quadrature points/weights (µm)
+  r_max:          50.0
+  # Number of quadrature points for aerosol radius
+  nquad_radius:   1000
+  # Reference wavelength (µm)
+  λ_ref:          0.55 
+  # Decomposition method (NAI2 or PCW)
+  decomp_type:        NAI2()
