### 2022-07-11 e3367d9
Fix AOD ref and final start/stop limits
