### 2022-07-11 d097856
Fix AOD ref and final start/stop limits
