### 2022-07-13 244eacc
Generalize viewing directions and types
