### 2022-07-13 378b358
Generalize viewing directions and types
