+++ test/rami/rami_v2.jl
@@ -2,14 +2,20 @@ using JSON
 using vSmartMOM
 using DelimitedFiles
 using Distributions
-
-# TODO: Add proper band limits and resolutions for gases:
-const sentinel_band_to_wn = Dict([("2" , 1e7 ./ reverse(collect(455.0:1:534.0))),    # Blue
-                                  ("3" , 1e7 ./ [559.8 559.9]),    # Green
-                                  ("4" , 1e7 ./ [664.6 664.7]),    # Red
-                                  ("8A", 1e7 ./ [864.7 864.8]),    # Narrow NIR 
-                                  ("11", 1e7 ./ [1613.7 1613.8]),  # SWIR
-                                  ("12", 1e7 ./ [2202.4 2202.5])]) # SWIR 
+                                  
+const sentinel_band_to_index = Dict([("2" , 1),    # Blue
+                                     ("3" , 2),    # Green
+                                     ("4" , 3),    # Red
+                                     ("8A", 4),    # Narrow NIR 
+                                     ("11", 5),  # SWIR
+                                     ("12", 6)]) # SWIR
+const scenario_yaml_I = Dict([("AtmosphereType.RAYLEIGH"           , "RamiNoGasI.yaml"),    
+                            ("AtmosphereType.ABSORBING"            , "RamiGasI.yaml"),      
+                            ("AtmosphereType.SCATTERING_ABSORBING" , "RamiGasI.yaml"),      
+                            ("AtmosphereType.AEROSOLS"             , "RamiNoGasI.yaml"),     
+                            ("AtmosphereType.SCATTERING_AEROSOLS"  , "RamiNoGasI.yaml"),    
+                            ("AtmosphereType.ABSORBING_AEROSOLS"   , "RamiGasI.yaml"),
+                            ("AtmosphereType.COMPLETE"             , "RamiGasI.yaml" )]) # SWIR
 
 
 ## 
@@ -28,8 +34,7 @@ US_std_atmos = readdlm(rami_atmos_fn, ' ')
 # Starts with HOM45_LAM, HOM25_LAM, or HOM35_LAM for homogeneous discrete with isotropic background
 
 function produce_rami_results(experiment_name::String;
-                              rami_json::String = "test/rami/RAMI4ATM_experiments_v1.0.json",
-                              default_params::String = "test/rami/RamiNoGas.yaml")
+                              rami_json::String = "test/rami/RAMI4ATM_experiments_v1.0.json")
 
     # Get rami scenario from experiment name
     all_scenarios = JSON.parsefile(rami_json);
@@ -37,91 +42,44 @@ function produce_rami_results(experiment_name::String;
     @assert length(all_scenarios) ≤ 1 "Multiple matching experiment names!"
     @assert length(all_scenarios) ≥ 1 "Experiment name not found in JSON input"
     curr_scenario = all_scenarios[1]["observations"][1];
-
-    # Get a default set of Rayleigh params
-    params = vSmartMOM.parameters_from_yaml(default_params)
-
-    # ########################################
-    # Change params fields to match rami input
-    # ########################################
+    band = curr_scenario["measures"][1]["bands"]
 
     # There are 7 rami categories to be mindful of: 
     # atmosphere, name, canopy, illumination, measures, time, surface
 
-    rami_atmosphere = curr_scenario["atmosphere"]
-    rami_name = curr_scenario["name"]
-    rami_canopy = curr_scenario["canopy"]
+    rami_atmosphere   = curr_scenario["atmosphere"]
+    rami_name         = curr_scenario["name"]
[…]
