### 2022-07-14 b732a33
Add bands, gas scaling and WL dependent refractive
