+++ test/rami/createAbscoRAMI.jl
@@ -1,19 +1,22 @@
+using vSmartMOM
 using vSmartMOM.Absorption
-using Unitful
 
+
+folder = "/net/fluo/data2/data/Rami_CS_database/"
 # Pressure and temperature grids
-pressures = 1e-10:25.0:1050
-temperatures = 180.0:5:340 
-Δλ = 0.1
+pressures = 1e-5:25.0:1050
+temperatures = 180.0:10:340 
+Δν = 0.001
 
-bands = ([1e7/456.0, 1e7/533.0],
-        [1e7/538.0, 1e7/583.0],
-        [1e7/646.0, 1e7/684.0],
-        [1e7/837.0, 1e7/881.0],
-        [1e7/1539.0, 1e7/1682.0],
-        [1e7/2078.0, 1e7/2320.0])
+bands = ([1e7/456.0,  1e7/533.0],
+         [1e7/538.0,  1e7/583.0],
+         [1e7/646.0,  1e7/684.0],
+         [1e7/837.0,  1e7/881.0],
+         [1e7/1539.0, 1e7/1682.0],
+         [1e7/2078.0, 1e7/2320.0])
 
 molecules = ("CO2", "H2O", "O3", "N2O", "CO", "CH4", "O2")
+
 
 for iMolec in eachindex(molecules)
     specFile = artifact(molecules[iMolec])
@@ -21,25 +24,46 @@ for iMolec in eachindex(molecules)
         start = minimum(bands[iBand])-1
         stop  = maximum(bands[iBand])+1
         @show specFile, start, stop
-        outFile = "rami_spectroscopy_" * molecules[iMolec] * "_" * string(iBand) * ".jld2"
+        outFile = folder * "rami_spectroscopy_" * molecules[iMolec] * "_" * string(iBand) * ".jld2"
         @show outFile
         try
             test_ht = Absorption.read_hitran(specFile, ν_min=start-20, ν_max=stop+20)
             
             # Wavelength grid
-            ν_grid = start:Δλ:stop
+            ν_grid = start:Δν:stop
             # Create the Interpolation Model
             interp_model = make_interpolation_model(test_ht, Voigt(), ν_grid, pressures, 
                                         temperatures, CEF=HumlicekWeidemann32SDErrorFunction()) 
-            save_interpolation_model(interp_model, )
-        catch 
-            println("No hitran?")
+            save_interpolation_model(interp_model, outFile)
+        catch e
+            println(e)
         end
     end
 end
 
-# Get the test data
-#CO2_file = artifact("CO2")
-#test_ht = Absorption.read_hitran(CO2_file, mol=2, iso=1, ν_min=6000, ν_max=6400)
+# Ozone separately: #####################
+using DelimitedFiles
+o3 = readdlm("//net/fluo/data2/data/Rami_CS_database/serdyuchenkogorshelev5digits.dat", skipstart=10000)
+ν_grid    = o3[1,1]:0.01:o3[end,1]
+# Remove <0 values:
+o3[o3.<0] .= 0
+cs_matrix = o3[:,end:-1:2]
+d1,d2 = size(cs_matrix)
+#cs_matrix = reshape(cs_matrix,(d1,1,d2))
+cs_matrix_2 = zeros(d1,2,d2);
+cs_matrix_2[:,1,:] = cs_matrix
+cs_matrix_2[:,2,:] = cs_matrix
 
-#grid = 6000:0.01:6400;
+# Create Interpolation model from Interpolations.jl:
+itp = interpolate(cs_matrix_2,  (BSpline(Linear()),BSpline(Constant()),BSpline(Linear())))
+# Use flat extrapolation:
+etpf = extrapolate(itp, Flat())
[…]
