### 2022-07-14 60e2177
Add bands, gas scaling and WL dependent refractive
