+++ test/rami/RamiNoGasI.yaml
@@ -67,22 +67,3 @@ atmospheric_profile:
 
   # Reduce profile to n layers
   profile_reduction: -1
-
-scattering:
-  # List of scattering aerosols and their properties
-  aerosols:       
-    - τ_ref:          0.2         # Reference τ
-      μ:              1.0           # effective radius (µm)
-      σ:              1.5           # geometric standard deviation (µm)
-      nᵣ:             1.3           # Real part of refractive index
-      nᵢ:             0.00000001    # Imag part of refractive index
-      p₀:             700.0         # Pressure peak (hPa)
-      σp:             50.0          # Pressure peak width (hPa)
-  # Maximum aerosol particle radius for quadrature points/weights (µm)
-  r_max:          50.0
-  # Number of quadrature points for aerosol radius
-  nquad_radius:   1000
-  # Reference wavelength (µm)
-  λ_ref:          0.55 
-  # Decomposition method (NAI2 or PCW)
-  decomp_type:        NAI2()

[…]
