### 2022-07-14 36c5e64
Turn off autodiff test for aerosols in test
