### 2022-07-18 #4575 @jimmielin
Hello all, I've had an opportunity to look at this package during the pre-review phase, and have just went through the software and manuscript again. Please find my comments below.

## Summary
`vSmartMOM.jl` is a Julia package for computing atmospheric optical properties and fully-polarized multiple-scattering radiance simulations with the Matrix Operator Method. The software is open-source licensed, is modular (including `Absorption.jl` and `Scattering.jl` and having separate documentation sections for these is highly positive, in my view, it shows that modularity was not an afterthought), and takes advantage of specific strengths of the Julia ecosystem, including GPU-acceleration and automatic differentiation. The development of a fully Julia-based package for remote sensing applications is also a good step in improving software capabilities for researchers. The manuscript is well written to showcase its strengths, and the documentation is clear and covers the functionality of the package. I recommend publication after the following comments are addressed:

## Functionality
- The RT parameters are well documented. However, I had to dig into the documentation then into the `DefaultParameters.yaml` to configure the `architecture` property for testing CPU and GPUs. Also, using `GPU()` and `CPU()` as specified in the yaml comments did not work out-of-the-box:
```
➜  julia  julia vSmartMOMtest.jl
[ Info: CUDA-enabled GPU(s) detected
[ Info: CUDA.CuDevice(0): NVIDIA GeForce RTX 3050 Ti Laptop GPU
WARNING: both KernelAbstractions and Architectures export "CPU"; uses of it in module CoreRT must be qualified
```
```yaml
...
  # Architecture (default_architecture, GPU(), CPU())
  architecture:       default_architecture
...
```
Simply noting `Architectures.GPU()`, `Architectures.CPU()` instead would help, so that the documentation is accurate out-of-the-box.

- The manuscript claims "optimized techniques have been implemented to speed up the package's performance on both CPU and GPU by orders-of-magnitude compared to existing radiative transfer codes" (L22-24). This claim is not further explained in the Benchmarks section, which only compares the CPU vs. GPU architecture in vSmartMOM.jl. I would be interested in an example of specific order-of-magnitude improvement compared to existing codes (e.g., SCIATRAN?).

## Documentation
- The project has a clear Code of Conduct which was added very early in the project, which I highly appreciate. There is also a statement about support and suggestions in the `README`. However, there is no clear statement of how to contribute to the package, and whether standard GitHub practices (Issues & Pull Requests) are supported or these should be sent by email (as mentioned in the "Support" section of the README). I would recommend the authors to clarify how contributions of the package are accepted, preferably through GitHub and not by email which is private and results in duplicate efforts to report bugs or contribute to the software.
- Docstrings are missing in the "Methods & Types" section of the documentation at https://remotesensingtools.github.io/vSmartMOM.jl/dev/pages/vSmartMOM/Types/

## Software paper
- MOM methods date back to at least Liu & Ruprecht, 1996. The authors cite Sanghavi et al., 2013 (L13, L46). While this is a software package paper which implements Sanghavi et al. 2013's vSmartMOM algorithm, I still recommend more references to the basis of the numerical methods. This will also help put this software package in context (what were the codes available before `vSmartMOM.jl`?)
- L61 first mentions CPU and GPU (CUDA) architectures. I suggest saying CPU core is single-threaded here first, along with L78.

## Technical Corrections
- Citations on line 57 are missing (Yurchenko et al., 2012; Rothman et al., 2010)
- The title of Hu et al., 2000 is incorrectly parsed (??-fit instead of $\delta$-fit)
