+++ test/rami/rami_tools.jl
@@ -1,3 +1,50 @@
+using Interpolations
+using DelimitedFiles
+using Printf
+
+
+# Read Spectral Response Functions:
+sr = readdlm("test/rami/Sentinel2A_ILS.dat")
+n_desert      = readdlm("test/rami/refractive_aero_desert.txt")
+n_continental = readdlm("test/rami/refractive_aero_continental.txt")
+
+const sr_2 = sr[:,3]
+const sr_3 = sr[:,4]
+const sr_4 = sr[:,5]
+const sr_8A = sr[:,10]
+const sr_11 = sr[:,13]
+const sr_12 = sr[:,14]
+const wl_sr = sr[:,1]
+
+# Some LUT for convenience:
+const sentinel_band_to_index = Dict([("2" , 1),    # Blue
+                                     ("3" , 2),    # Green
+                                     ("4" , 3),    # Red
+                                     ("8A", 4),    # Narrow NIR 
+                                     ("11", 5),  # SWIR
+                                     ("12", 6)]) # SWIR
+
+const scenario_yaml_I = Dict([("AtmosphereType.RAYLEIGH"           , "RamiNoGasI.yaml"),    
+                            ("AtmosphereType.ABSORBING"            , "RamiGasI.yaml"),      
+                            ("AtmosphereType.SCATTERING_ABSORBING" , "RamiGasI.yaml"),      
+                            ("AtmosphereType.AEROSOLS"             , "RamiNoGasI.yaml"),     
+                            ("AtmosphereType.SCATTERING_AEROSOLS"  , "RamiNoGasI.yaml"),    
+                            ("AtmosphereType.ABSORBING_AEROSOLS"   , "RamiGasI.yaml"),
+                            ("AtmosphereType.COMPLETE"             , "RamiGasI.yaml" )]) # SWIR
+
+const noRayleighType   = ("AtmosphereType.ABSORBING", "AtmosphereType.AEROSOLS", "AtmosphereType.ABSORBING_AEROSOLS")
+const noAbsorptionType = ("AtmosphereType.RAYLEIGH", "AtmosphereType.AEROSOLS", "AtmosphereType.SCATTERING_AEROSOLS")
+
+
+const sentinel_band_to_LUT = Dict([("2" , 2),    # Blue
+                            ("3" , 3),    # Green
+                            ("4" , 4),    # Red
+                            ("8A", 9),    # Narrow NIR 
+                            ("11", 12),  # SWIR
+                            ("12", 13)]) # SWIR
+
+
+# Add aerosols (if required by RAMI, otherwise do nothing)
 function add_aerosols!(rami_atmosphere, params, band, n_desert, n_continental)# TODO We still need to check the two fractions of coarse and fine! Right now, only one is
     if length(rami_atmosphere["aerosols"]) > 0# 
         if startswith(rami_atmosphere["aerosols"][1]["name"], "D")
@@ -18,11 +65,12 @@ function add_aerosols!(rami_atmosphere, params, band, n_desert, n_continental)# 
             # TODO: Make it wavelength dependent (and reference scenario fixed)
             #nᵣ = 1.477538814814815
             #nᵢ = 0.004342592592592592
-            @show band[1]
-            iN = sentinel_band_to_LUT[band[1]]
+            @show band
+            iN = sentinel_band_to_LUT[band]
             nᵣ = n_desert[iN,2]
             #nᵣ = 1.4220380000000001 
             nᵢ = n_desert[iN,3]
+            n_ref = n_desert[3,2] - im*n_desert[3,3]
[…]
