+++ test/rami/rami_v2.jl
@@ -3,148 +3,54 @@ using vSmartMOM
 using DelimitedFiles
 using Distributions
 
-# Read Spectral Response Functions:
-sr = readdlm("test/rami/Sentinel2A_ILS.dat")
-n_desert      = readdlm("test/rami/refractive_aero_desert.txt")
-n_continental = readdlm("test/rami/refractive_aero_continental.txt")
-sr_2 = sr[:,3]
-sr_3 = sr[:,4]
-sr_4 = sr[:,5]
-sr_8A = sr[:,10]
-sr_11 = sr[:,13]
-sr_12 = sr[:,14]
-wl_sr = sr[:,1]
-
-# Some LUT for convenience:
-const sentinel_band_to_index = Dict([("2" , 1),    # Blue
-                                     ("3" , 2),    # Green
-                                     ("4" , 3),    # Red
-                                     ("8A", 4),    # Narrow NIR 
-                                     ("11", 5),  # SWIR
-                                     ("12", 6)]) # SWIR
-
-const scenario_yaml_I = Dict([("AtmosphereType.RAYLEIGH"           , "RamiNoGasI.yaml"),    
-                            ("AtmosphereType.ABSORBING"            , "RamiGasI.yaml"),      
-                            ("AtmosphereType.SCATTERING_ABSORBING" , "RamiGasI.yaml"),      
-                            ("AtmosphereType.AEROSOLS"             , "RamiNoGasI.yaml"),     
-                            ("AtmosphereType.SCATTERING_AEROSOLS"  , "RamiNoGasI.yaml"),    
-                            ("AtmosphereType.ABSORBING_AEROSOLS"   , "RamiGasI.yaml"),
-                            ("AtmosphereType.COMPLETE"             , "RamiGasI.yaml" )]) # SWIR
-
-const noRayleighType = ("AtmosphereType.ABSORBING", "AtmosphereType.AEROSOLS", "AtmosphereType.ABSORBING_AEROSOLS")
-
-const sentinel_band_to_LUT = Dict([("2" , 2),    # Blue
-                            ("3" , 3),    # Green
-                            ("4" , 4),    # Red
-                            ("8A", 9),    # Narrow NIR 
-                            ("11", 12),  # SWIR
-                            ("12", 13)]) # SWIR
-## 
+include("rami_tools.jl")
 
 # Read in the standard atmosphere
 
 # atmosphere ascii file location (MAYBE add as artifact)
-rami_atmos_fn = "test/rami/RAMI4ATM_AFGLUSstandard_ap_v1.0.txt"
-US_std_atmos = readdlm(rami_atmos_fn, ' ')
-
-
-## 
+# rami_atmos_fn = "test/rami/RAMI4ATM_AFGLUSstandard_ap_v1.0.txt"
+# US_std_atmos = readdlm(rami_atmos_fn, ' ')
 
 # Starts with HOM00_LAM, HOM00_BLA, or HOM00_WHI for lambertian surfaces
 # Starts with HOM00_RPV, HOM00_RLI for Anisotropic surfaces
 # Starts with HOM45_LAM, HOM25_LAM, or HOM35_LAM for homogeneous discrete with isotropic background
 
 function produce_rami_results(experiment_name::String;
-                              rami_json::String = "test/rami/RAMI4ATM_experiments_v1.0.json",
-                              n_desert=n_desert,n_continental=n_continental, sr=sr )
+                        all_scenarios = all_scenarios,
+                        n_desert=n_desert,n_continental=n_continental, sr=sr )
 
     # Get rami scenario from experiment name
-    all_scenarios = JSON.parsefile(rami_json);
-    filter!(x->x["name"] == experiment_name, all_scenarios);
-    @assert length(all_scenarios) ≤ 1 "Multiple matching experiment names!"
[…]
