+++ test/benchmarks/prototype_inelastic_VS2.jl
@@ -1,0 +1,214 @@
+using Revise
+using vSmartMOM, vSmartMOM.CoreRT, vSmartMOM.SolarModel
+using vSmartMOM.InelasticScattering
+using Statistics
+using Interpolations
+using InstrumentOperator #for convolution of hires spectrum to instrument grid
+using ImageFiltering
+using Distributions
+using Plots
+
+Fraunhofer=true
+λ₀ = 400. #745. #400. #nm
+#iBand = 1
+n2,o2 = InelasticScattering.getRamanAtmoConstants(1.7/λ₀, 300.);
+FT = Float64
+RS_type = InelasticScattering.VS_0to1_plus{FT}(
+                   n2=n2,
+                   o2=o2);
+# Load YAML files into parameter struct
+parameters = parameters_from_yaml("test/test_parameters/O2ParametersVS.yaml");
+# Create model struct (precomputes optical properties) from parameters
+model      = model_from_parameters(RS_type, λ₀, parameters);
+
+if Fraunhofer
+    T_sun = 5777. # KRS 
+    Tsolar = solar_transmission_from_file("/home/sanghavi/code/github/vSmartMOM.jl/src/SolarModel/solar.out")
+    Tsolar_interp = LinearInterpolation(Tsolar[4:end, 1], Tsolar[4:end, 2])
+    RS_type.F₀ = zeros(model.params.polarization_type.n, 
+                        sum(length(RS_type.grid_in[i]) for i in 1:length(RS_type.iBand)))
+    local i_offset = 0
+    for iB=1:length(RS_type.iBand)
+        ν = collect(RS_type.grid_in[iB])
+        P = planck_spectrum_wn(T_sun, ν)    
+        F₀ = zeros(length(ν));
+        
+        for i=1:length(ν)
+            sol_trans = Tsolar_interp(ν[i]);
+            F₀[i] = sol_trans * P[i];
+            RS_type.F₀[1,i+i_offset] = F₀[i];
+        end 
+        i_offset += length(ν)
+    end
+else
+    RS_type.F₀ = zeros(model.params.polarization_type.n, 
+        sum(length(RS_type.grid_in[i]) for i in 1:length(RS_type.iBand)))
+    local i_offset = 0
+    for iB=1:length(RS_type.iBand)
+        ν = collect(RS_type.grid_in[iB])
+
+        for i=1:length(ν)
+            RS_type.F₀[1,i+i_offset] = 1.0;
+        end 
+        i_offset += length(ν)
+    end
+end
+
+R, T, ieR, ieT = CoreRT.rt_run_test(RS_type,
+           model, RS_type.iBand);
+
+############# RRS ##############
+parameters2 = 
+       parameters_from_yaml("test/test_parameters/FraunhoferMockParameters.yaml");
+#parameters = parameters_from_yaml("test/test_parameters/O2Parameters2.yaml");
[…]
