+++ src/Inelastic/raman_atmo_prop.jl
@@ -59,7 +59,7 @@ function getRamanSSProp!(RS_type::RRS, λ, grid_in)
     #n2, o2 = getRamanAtmoConstants(1.e7/λ, T)
     # determine Rayleigh scattering cross-section at central wavelength λ of the spectral band (assumed constant throughout the band)
     atmo_σ_Rayl = compute_optical_Rayl(λ, n2, o2)
-    greek_raman = get_greek_raman(RS_type, n2, o2)
+    RS_type.greek_raman = get_greek_raman(RS_type, n2, o2)
     #get_greek_raman!(RS_type, n2, o2)
     RS_type.ϖ_Cabannes .= compute_ϖ_Cabannes(RS_type, λ, n2, o2)
     # @show RS_type.ϖ_Cabannes
@@ -143,7 +143,7 @@ function getRamanSSProp!(
     nBand = 3; #for n2 and o2 each
     iBand = [1,2,3];
     ν̄ = nm_per_m/λ_inc
-    push!(grid_in, ν̄:0.1:ν̄)
+    push!(grid_in, ν̄:0.3:ν̄)
     molec = [n2,o2]
     #========================================================================#
     for mol in molec
@@ -159,7 +159,7 @@ function getRamanSSProp!(
         band_max += nm_per_m/λ_inc;
         band_max += 2.;
 
-        push!(grid_in, band_min:0.05:band_max)
+        push!(grid_in, band_min:0.3:band_max)
 
     end
     #========================================================================#
@@ -179,7 +179,7 @@ function getRamanSSProp!(
         
         if iB==1
             ϖ_Cabannes[iB] = InelasticScattering.compute_ϖ_Cabannes(RS_type, λ_inc, n2, o2)
-            @show ϖ_Cabannes
+            #@show ϖ_Cabannes
         else
             ϖ_Cabannes[iB] = 1.
         end
@@ -256,4 +256,147 @@ function getRamanSSProp!(
             greek_raman_VS_o2
     return nothing
 end
+
+function getRamanSSProp!(
+    RS_type::VS_1to0_plus, λ_inc)
+
+@unpack n2,o2,
+    iBand, grid_in, bandSpecLim, 
+    greek_raman, greek_raman_VS_n2, greek_raman_VS_o2,
+    ϖ_Cabannes, fscattRayl,
+    ϖ_λ₁λ₀, i_λ₁λ₀,
+    ϖ_λ₁λ₀_VS_n2, i_λ₁λ₀_VS_n2,
+    ϖ_λ₁λ₀_VS_o2, i_λ₁λ₀_VS_o2,
+    i_λ₁λ₀_all =  RS_type
+#n2, o2 = getRamanAtmoConstants(1.e7/λ, T)
+iBand = []
+grid_in = []
+bandSpecLim = []
+
+nm_per_m = 1.e7;
+greek_raman = get_greek_raman(RS_type, n2, o2)
+greek_raman_VS_n2 = get_greek_raman_VS(RS_type, n2)
+greek_raman_VS_o2 = get_greek_raman_VS(RS_type, o2)
+#λ_scatt = 2.e7/(grid_in[1]+grid_in[end])     
[…]
