+++ src/CoreRT/CoreKernel/doubling_inelastic.jl
@@ -196,7 +196,7 @@ function doubling_helper!(RS_type::Union{VS_0to1_plus, VS_1to0_plus},
             tmp1 = gp_refl ⊠  (J₀⁺ + r⁻⁺ ⊠ J₁⁻)
             tmp2 = gp_refl ⊠  (J₁⁻ + r⁻⁺ ⊠ J₀⁺)
             #for n₁ in eachindex ieJ₁⁺[1,1,:,1]
-            for Δn in length(i_λ₁λ₀_all)
+            for Δn = 1:length(i_λ₁λ₀_all)
                 n₁ = i_λ₁λ₀_all[Δn]
                 n₀ = 1
                 if n₁>0
@@ -307,11 +307,12 @@ end
     if 1 ≤ n₀ ≤ size(ier⁻⁺,4)
         i = mod(iμ, n_stokes)
         j = mod(jμ, n_stokes)
-        if (i > 2)
+        if !(1<=i<=2) #(i > 2)
             ier⁻⁺[iμ,jμ,n,n₀] = - ier⁻⁺[iμ, jμ, n, n₀]
         end
         
-        if ((i <= 2) & (j <= 2)) | ((i > 2) & (j > 2))
+        #if ((i <= 2) & (j <= 2)) | ((i > 2) & (j > 2))
+        if (((1<=i<=2) & (1<=j<=2)) | (!(1<=i<=2) & !(1<=j<=2)))
             ier⁺⁻[iμ,jμ,n,n₀] = ier⁻⁺[iμ,jμ,n,n₀]
             iet⁻⁻[iμ,jμ,n,n₀] = iet⁺⁺[iμ,jμ,n,n₀]
         else
@@ -325,22 +326,24 @@ end
                         ier⁻⁺, iet⁺⁺, ier⁺⁻, iet⁻⁻)
     iμ, jμ, Δn  = @index(Global, NTuple)
     #@unpack i_λ₁λ₀ = RS_type 
+    #@show "here 3.1"
     n  = i_λ₁λ₀_all[Δn]
     i = mod(iμ, n_stokes)
     j = mod(jμ, n_stokes)
-
-    if (i > 2)
-        ier⁻⁺[iμ,jμ,n,1] = - ier⁻⁺[iμ, jμ, n, 1]
+    #@show n, i, j
+    if (n>0)
+        if !(1<=i<=2)
+            ier⁻⁺[iμ,jμ,n,1] = - ier⁻⁺[iμ, jμ, n, 1]
+        end
+        
+        if ((1<=i<=2) & (1<=j<=2)) | (!(1<=i<=2) & !(1<=j<=2))
+            ier⁺⁻[iμ,jμ,n,1] = ier⁻⁺[iμ,jμ,n,1]
+            iet⁻⁻[iμ,jμ,n,1] = iet⁺⁺[iμ,jμ,n,1]
+        else
+            ier⁺⁻[iμ,jμ,n,1] = - ier⁻⁺[iμ,jμ,n,1]
+            iet⁻⁻[iμ,jμ,n,1] = - iet⁺⁺[iμ,jμ,n,1]
+        end
     end
-    
-    if ((i <= 2) & (j <= 2)) | ((i > 2) & (j > 2))
-        ier⁺⁻[iμ,jμ,n,1] = ier⁻⁺[iμ,jμ,n,1]
-        iet⁻⁻[iμ,jμ,n,1] = iet⁺⁺[iμ,jμ,n,1]
-    else
-        ier⁺⁻[iμ,jμ,n,1] = - ier⁻⁺[iμ,jμ,n,1]
-        iet⁻⁻[iμ,jμ,n,1] = - iet⁺⁺[iμ,jμ,n,1]
[…]

[…]
