+++ test/rami/rami_v2.jl
@@ -3,7 +3,7 @@ using vSmartMOM
 using DelimitedFiles
 using Distributions
 
-include("rami_tools.jl")
+include("test/rami/rami_tools.jl")
 
 # Read in the standard atmosphere
 
@@ -75,7 +75,8 @@ all_scenarios = JSON.parsefile(rami_json);
 # R, models = produce_rami_results("HOM00_WHI_SD2S_M02_z30a000")
 #R, models = produce_rami_results("HOM00_LAM_ED2D_M12_z30a000")
 #R, models = produce_rami_results("HOM00_WHI_A00S_M02_z30a000")
-BRF, R, model = produce_rami_results("HOM00_WHI_A00S_M12_z30a000")
+#BRF, R, model = produce_rami_results("HOM00_WHI_A00S_M12_z30a000")
+BRF, R, model = produce_rami_results("HOM00_WHI_SD2S_M02_z30a000")
 
 
 ##
