### 2022-07-24 RemoteSensingTools/vSmartMOM.jl#127 @arjunsavel
JOSS review: Paper edits
**Describe the bug**
There are a few small edits to the manuscript for this project's corresponding JOSS paper that I'd suggest.

- like brown dwarfs and exoplanets —> such as brown dwarfs and exoplanets
- fully-polarized —> fully polarized
- flexible and —> flexible, and
- orders-of-magnitude —> orders of magnitude
- barrier-of-entry —> barrier of entry
- redesigns the radiative transfer code —> designs a radiative transfer code
- like GPU —> such as GPU
- Perhaps describe why GPU acceleration and autodiff are beneficial for retrievals
- high fidelity —> high-fidelity
- Was RT defined?
- 0.01 step size —>  0.01 cm^-1 step size
- Are spectra obtained from ExoMOL and HITEMP, or are the opacity data obtained from them, using HITRAN by default otherwise?
- The Q and U benchmark comparisons should also be stated in percent, for consistency (and to make it clear that it's a fractional comparison).
- Define auto-differentiation
- Like delta truncation —> such as delta truncation
- Note that the 100x speedup is only for one of the simulations (10^5 spectral points), as opposed to lower numbers of spectral points.
- performance optimizations in vSmartMOM.jl suggest —>  performance optimizations in vSmartMOM.jl, suggest
- Could the comparison to literature tables be presented as a plot, as well? It's a bit tricky to absorb in a table.
- Nice benchmark to other codes! It’d also be useful to briefly *describe* other codes and differences. e.g., do all radiative transfer codes use line-by-line calculations?
