### 2022-07-24 RemoteSensingTools/vSmartMOM.jl#128 @arjunsavel
JOSS review: Documentation edits
**Describe the bug**
Below are some small edits to this code's documentation, as part of this project's [JOSS review.](https://github.com/openjournals/joss-reviews/issues/4575)

*Overall comments*
- I believe the internal links aren't working, e.g. the example page link in the [Absorption Module Overview.](https://remotesensingtools.github.io/vSmartMOM.jl/dev/pages/Absorption/Overview/)

*Getting started page*
- computationally-intensive —> computationally intensive
- easy-to-use —> easy to use
- at specified pressure —> over specified pressure, perhaps?
-  ] and —> ], and
- as seen on the left navigation bar —> navigation bar (because at the top for smaller screens)
- guide to the all the —> guide to all the

*vSmartMOM module overview*
- It's implied that this is just for the Earth's atmosphere, but the corresponding paper mentions exoplanet and brown dwarf atmospheres. Perhaps a broader statement could be made here?
- Link matrix operator method description somewhere?
- high fidelity —> high-fidelity
- Excellent diagrams here and elsewhere, by the way!

*RT parameters guide*
- depol is unitless, correct? It might be helpful to note this (and other unitless quantities) as such.
- imag —> imaginary
- Under aerosols, the units don't have the same bolding as above.

*Scattering mod overview*
- computationally-intensive —> computationally intensive

*Absorption tutorial*
- lineshapes or line shapes? I believe there's inconsistent usage.
- And how the —> , and how the
- tutorials here —> tutorials here.
- but we are —> , but we are
- kine-mixing, collision —> line-mixing, and collision-induced
- and we will walk —>, and we will walk
- here now —> here
- doppler shifted —> doppler-shifted
- then but let’s —> the, but let’s
- A few more words about how to get the Doppler width for broadening? Noting that it’s the expectation value or something?
- A few of the markdown list items aren't rendered correctly, and instead still have the <li> tags
- to get FWHM —> to get the full-width half-max (FWHM) of the spectral line.
- but this is mostly —> , but the is mostly
- It'd be helpful to describe the variables used in the tutorial's equations.
- Pressure broadening hasn’t been defined; perhaps it'd be helpful to do so before its usage.
- See, for instance, here —> See, for instance, here.
- Describe what’s happening in the plots — what do we expect? What do we see?
- Does that hitranmodel cell need to be there?

*Scattering tutorial*
- are Greek coefficients capitalized?
- is Legendre capitalized?
- As with the previous tutorial: Describe what’s happening in the plots — what do we expect? What do we see?
