### 2022-07-24 RemoteSensingTools/vSmartMOM.jl#129 @arjunsavel
JOSS review: Errors in absorption and scattering examples with make_mie_model?
**Describe the bug**
When running the [absorption](https://remotesensingtools.github.io/vSmartMOM.jl/dev/pages/Absorption/Example/) and [scattering](https://remotesensingtools.github.io/vSmartMOM.jl/dev/pages/Scattering/Example/) examples, I ran into an issue with `make_mie_model`. It seems like there might be a type error when making the function call.

**To Reproduce**
Steps to reproduce the behavior:
1. Run the [absorption tutorial.](https://remotesensingtools.github.io/vSmartMOM.jl/dev/pages/Scattering/Example/)
2. The code should break on `model_PCW_computed_wigner = make_mie_model(PCW(), aero, λ, polarization_type, truncation_type, wigner_A, wigner_B)`

**Expected behavior**
I thought the example code would run with my current dependency setup, but there's of course the possibility that I have incorrect package versions! Happy to provide more info if helpful.

**Screenshots**

 - OS: MacOS Monterey
 - Version 12.2.1
