### 2022-07-24 #4575 @arjunsavel
Hi all, I've finished my look at this package!

With `vSmartMOM`, the authors have put together a well-developed package and a welcome addition to the open-source radiative transfer literature. The code is fast, modular, and appears well-tested, and it definitely has sufficient documentation and tutorials. I have a few comments, and pending their resolution I recommend this project for acceptance.

Firstly, I have some minor paper and documentation comments, which are grouped in issues in the code repository:

- [Paper comments](https://github.com/RemoteSensingTools/vSmartMOM.jl/issues/127)
- [Documentation comments](https://github.com/RemoteSensingTools/vSmartMOM.jl/issues/128)
- [Documentation example bug?](https://github.com/RemoteSensingTools/vSmartMOM.jl/issues/129)

Beyond these issues, I was curious about the authors' thoughts on the code's [demos.](https://github.com/RemoteSensingTools/vSmartMOM.jl/tree/main/demo) They appear to constitute a potentially valuable set of additional examples and use cases, but there are some obstacles to their immediate application. For instance, they reference what I surmise to be a past name for `vSmartMOM` (`RadiativeTransfer`). Additionally, certain modules (`RadiativeTransfer.PhaseFunction`) and methods (`make_univariate_aerosol`) appear to be deprecated. If the authors plan to continue packaging these demos with the code, I recommend that they be updated to be executable with the current code version.
