### 2022-08-10 #4575 @cfranken
> Beyond these issues, I was curious about the authors' thoughts on the code's [demos.](https://github.com/RemoteSensingTools/vSmartMOM.jl/tree/main/demo) They appear to constitute a potentially valuable set of additional examples and use cases, but there are some obstacles to their immediate application. For instance, they reference what I surmise to be a past name for `vSmartMOM` (`RadiativeTransfer`). Additionally, certain modules (`RadiativeTransfer.PhaseFunction`) and methods (`make_univariate_aerosol`) appear to be deprecated. If the authors plan to continue packaging these demos with the code, I recommend that they be updated to be executable with the current code version.

Apologies for the late replies, we just started going through everything. As for demos, we saw now that they are indeed deprecated. It is likely best to remove the demos from that specific folder and make them part of the documentation, so that they are run automatically with every push, allowing us to see deprecation issues immediately. Thanks for catching that. We will go through the remaining issues soon as well.
