+++ src/CoreRT/Surfaces/rpv_surface.jl
@@ -17,7 +17,7 @@ Computes (in place) surface optical properties for a (scalar) lambertian albedo 
     - `τ_sum` total optical thickness from TOA to the surface
     - `architecture` Compute architecture (GPU,CPU)
 """ 
-function create_surface_layer!(rpv::rpvSurfaceScalar{FT}, 
+function create_surface_layer!(brdf::AbstractSurfaceType, 
                                added_layer::Union{AddedLayer,AddedLayerRS},
                                SFI,
                                m::Int,
@@ -34,11 +34,11 @@ function create_surface_layer!(rpv::rpvSurfaceScalar{FT},
     T_surf = arr_type(Diagonal(tmp))
     
     # Albedo normalized by π (and factor 2 for 0th Fourier Moment)
-    @show rpv
+    @show brdf
     if m==0
-        ρ = 2*vSmartMOM.CoreRT.reflectance(rpv, pol_type, Array(qp_μ), m)
+        @time ρ = 2*vSmartMOM.CoreRT.reflectance(brdf, pol_type, Array(qp_μ), m)
     else
-        ρ = vSmartMOM.CoreRT.reflectance(rpv, pol_type, Array(qp_μ), m)
+        @time ρ = vSmartMOM.CoreRT.reflectance(brdf, pol_type, Array(qp_μ), m)
     end
     
     # Move to architecture:
@@ -66,14 +66,18 @@ function create_surface_layer!(rpv::rpvSurfaceScalar{FT},
 end
 
 #Rahman Pinty Verstraete model
-function reflectance(rpv::rpvSurfaceScalar{FT}, μᵢ::FT, μᵣ::FT, dϕ::FT) where FT
+function reflectance(rpv::rpvSurfaceScalar{FT},  n, μᵢ::FT, μᵣ::FT, dϕ::FT) where FT
     @unpack ρ₀, ρ_c, k, Θ = rpv
     # TODO: Suniti, stupid calculations here:
-    θᵢ   = acos(μᵢ) #assert 0<=θᵢ<=π/2
-    θᵣ   = acos(μᵣ) #assert 0<=θᵣ<=π/2
-    cosg = -μᵢ*μᵣ + sin(θᵢ)*sin(θᵣ)*cos(dϕ) #RAMI form: μᵢ*μᵣ + sin(θᵢ)*sin(θᵣ)*cos(dϕ) (vSmartMOM sign convention is compatible with that of Rahman, Pinty, Verstraete, 1993) 
-    G    = (tan(θᵢ)^2 + tan(θᵣ)^2 + 2*tan(θᵢ)*tan(θᵣ)*cos(dϕ))^FT(0.5) #RAMI form: (tan(θᵢ)^2 + tan(θᵣ)^2 - 2*tan(θᵢ)*tan(θᵣ)*cos(dϕ))^FT(0.5)
-    return ρ₀ * rpvM(μᵢ, μᵣ, k) * rpvF(Θ, cosg) * rpvH(ρ_c, G)
+    if n==1
+        θᵢ   = acos(μᵢ) #assert 0<=θᵢ<=π/2
+        θᵣ   = acos(μᵣ) #assert 0<=θᵣ<=π/2
[…]
