+++ paper.bib
@@ -175,3 +175,33 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021ApJ...907...30S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article{grant1969discrete,
+  title={Discrete space theory of radiative transfer I. Fundamentals},
+  author={Grant, Ian Philip and Hunt, GE},
+  journal={Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences},
+  volume={313},
+  number={1513},
+  pages={183--197},
+  year={1969},
+  publisher={The Royal Society London}
+}
+@article{grant1969discrete,
+  title={Discrete space theory of radiative transfer II. Stability and non-negativity},
+  author={Grant, Ian Philip and Hunt, GE},
+  journal={Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences},
+  volume={313},
+  number={1513},
+  pages={199--216},
+  year={1969},
+  publisher={The Royal Society London}
+}
+@article{hunt1969discrete,
+  title={Discrete space theory of radiative transfer and its application to problems in planetary atmospheres},
+  author={Hunt, GE and Grant, IP},
+  journal={Journal of Atmospheric Sciences},
+  volume={26},
+  number={5},
+  pages={963--972},
+  year={1969}
+}
