### 2022-09-02 818ab70
Revise for exomol/hitemp
