+++ paper.md
@@ -56,7 +56,7 @@ The package has a modular architecture, allowing users to import just the specif
 
 ![Sample absorption spectrum of CO2 with 0.01 step size resolution, calculated using Absorption.jl](joss_2.png)
 
-**Absorption.jl** enables absorption cross-section calculations of atmospheric gases at different pressures, temperatures, and wavelengths. It uses the HITRAN [@Gordon:2017] database for calculations. For exoplanets and brown dwarfs, spectra are obtained from the ExoMOL/HITEMP databases [@yurchenko2012exomol:2012; @rothman2010hitemp:2010]. While it enables lineshape calculations from scratch, the module also allows users to create and save an interpolator object at specified wavelength, pressure, and temperature grids. The module also supports auto-differentiation of the profile, with respect to pressure and temperature. Calculations can be computed either on CPU or GPU (CUDA).
+**Absorption.jl** enables absorption cross-section calculations of atmospheric gases at different pressures, temperatures, and wavelengths. It uses the HITRAN [@Gordon:2017] database for calculations. For exoplanets and brown dwarfs, developments are underway to obtain spectra from the ExoMOL/HITEMP databases [@yurchenko2012exomol:2012; @rothman2010hitemp:2010], as described in @Sanghavi:2019. While the module enables lineshape calculations from scratch, it also allows users to create and save an interpolator object at specified wavelength, pressure, and temperature grids. The module also supports auto-differentiation of the profile, with respect to pressure and temperature. Calculations can be computed either on CPU or GPU (CUDA).
 
 ![Sample scattering phase functions of aerosols (I $\rightarrow$ I and the I $\rightarrow$ Q transition), calculated using Scattering.jl ($r_m$ = 10.0 $\mu$m, $\sigma$ = 1.1, $n_r$ = 1.3, $n_i$ = 0.0, $\lambda$ = 0.65 $\mu$m). Positive values are <span style="color:green">*green*</span>, and negative values are <span style="color:red">*red*</span>. ](joss_3.png)
 
