### 2022-09-02 9b89150
Refer to Grant & Hunt  papers
