+++ paper.bib
@@ -13,7 +13,7 @@
     year = {2013},
     month = {09},
     pages = {},
-    title = {VSmartMOM: A vector matrix operator method-based radiative transfer model linearized with respect to aerosol properties},
+    title = {vSmartMOM: A vector matrix operator method-based radiative transfer model linearized with respect to aerosol properties},
     volume = {133},
     journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
     doi = {10.1016/j.jqsrt.2013.09.004}
@@ -176,7 +176,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{grant1969discrete,
+@article{grant1969adiscrete,
   title={Discrete space theory of radiative transfer I. Fundamentals},
   author={Grant, Ian Philip and Hunt, GE},
   journal={Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences},
@@ -187,7 +187,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   publisher={The Royal Society London}
 }
 
-@article{grant1969discrete,
+@article{grant1969bdiscrete,
   title={Discrete space theory of radiative transfer II. Stability and non-negativity},
   author={Grant, Ian Philip and Hunt, GE},
   journal={Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences},
