### 2022-09-03 1e97705
Fix convolution
