+++ test/rami/rami_tools.jl
@@ -139,17 +139,40 @@ function convolve_2_sentinel(ν, BRF, band)
     ex = Meta.parse("sentinelILS = sr_" * band)
     eval(ex)
     srr = LinearInterpolation(wl_sr, sentinelILS);
-    @show size(BRF)
-    d1,d2,d3 = size(BRF)
+    si = size(BRF)
+    #@show size(BRF)
+    d1 = si[1]
 
     wl_in = 1e7./ν;
     weight = srr.(wl_in)
     weight = weight/sum(weight)
     BRF_out = zeros(d1)
     for i in eachindex(BRF_out)
+
         BRF_out[i] = weight' * BRF[i,1,:]
     end
     return BRF_out
+end
+
+# Convolve with Sentinel response function (input is in wavenumbers, convolution in nm space)
+function convolve_2_sentinel_HDR(ν, HDRF,BOAup,BOAdw, band)
+    ex = Meta.parse("sentinelILS = sr_" * band)
+    eval(ex)
+    srr = LinearInterpolation(wl_sr, sentinelILS);
+    si = size(HDRF)
+    #@show size(BRF)
+    d1 = si[1]
+
+    wl_in = 1e7./ν;
+    weight = srr.(wl_in)
+    weight = weight/sum(weight)
+    HDRF_out = zeros(d1)
+
+    for i in eachindex(HDRF_out)
+        HDRF_out[i] = weight' * (HDRF[i,1,:]./BOAdw)
+    end
+    BHR = (weight' * BOAup) / (weight' * BOAdw)
+    return HDRF_out, BHR
 end
 
 function getParams(scenario)
@@ -226,7 +249,7 @@ end
  
 
 function save_toa_results(BRF, model, name, folder)
-    fileNamePP = folder * name * "-brfpp_" * "vSmartMOM.mes" 
+    fileNamePP = folder * name * "-brfpp_" * "vSmartMOM-JPL.mes" 
     open(fileNamePP, "w") do io
         @printf io "%4d    %4d   %.6f \n" 76	5	-1.000000
         for i=1:76
@@ -234,7 +257,7 @@ function save_toa_results(BRF, model, name, folder)
         end
     end;
 
-    fileNameOP = folder * name * "-brfop_" * "vSmartMOM.mes" 
+    fileNameOP = folder * name * "-brfop_" * "vSmartMOM-JPL.mes" 
     open(fileNameOP, "w") do io
         @printf io "%4d    %4d   %.6f \n" 76	5	-1.000000
         for i=77:152
@@ -243,7 +266,84 @@ function save_toa_results(BRF, model, name, folder)
     end;
 end
 
+function save_boa_results(HDRF,BHR, model, name, folder)
+    fileNamePP = folder * name * "-hdrfpp_" * "vSmartMOM-JPL.mes" 
+    open(fileNamePP, "w") do io
+        @printf io "%4d    %4d   %.6f \n" 76	5	-1.000000
+        for i=1:76
+            @printf io "%.6f    %.6f   %.6f     %.6f    %.6f \n" deg2rad(model.obs_geom.sza) deg2rad(model.obs_geom.vza[i]) deg2rad(abs(model.obs_geom.vaz[i]-180.0)) HDRF[i] -1
+        end
+    end;
 
+    fileNameOP = folder * name * "-hdrfop_" * "vSmartMOM-JPL.mes" 
+    open(fileNameOP, "w") do io
+        @printf io "%.6f   %.6f \n" BHR	-1.000000
+    end;
+
+    fileNameBHR = folder * name * "-bhr_" * "vSmartMOM-JPL.mes" 
+end
+
+
+function produce_rami_results(experiment_name::String;
+    all_scenarios = all_scenarios,
[…]
