+++ src/CoreRT/Surfaces/rossli_surface.jl
@@ -1,5 +1,7 @@
 function reflectance(RossLi::RossLiSurfaceScalar{FT},n, μᵢ::FT, μᵣ::FT, dϕ::FT) where FT
     @unpack fiso, fvol, fgeo = RossLi
+    # Function was defined for RAMI definition, have to reverse here:
+    dϕ = π - dϕ
     # TODO: Suniti, stupid calculations here:
     θᵢ   = acos(μᵢ) #assert 0<=θᵢ<=π/2
     θᵣ   = acos(μᵣ) #assert 0<=θᵣ<=π/2
