+++ src/CoreRT/rt_run.jl
@@ -200,9 +200,9 @@ function rt_run(RS_type::AbstractRamanType,
 
     # Return R_SFI or R, depending on the flag
     #if RAMI
-    @show size(hdr), size(bhr_dw)
-    hdr = hdr ./ bhr_dw
-    return SFI ? (R_SFI, T_SFI, ieR_SFI, ieT_SFI, hdr, bhr_uw, bhr_dw) : (R, T)
+    #@show size(hdr), size(bhr_dw)
+    #hdr = hdr[:,1,:] ./ bhr_dw[1,:]
+    return SFI ? (R_SFI, T_SFI, ieR_SFI, ieT_SFI, hdr, bhr_uw[1,:], bhr_dw[1,:]) : (R, T)
     #else
     #return SFI ? (R_SFI, T_SFI, ieR_SFI, ieT_SFI) : (R, T)
     #end
