+++ src/CoreRT/rt_run.jl
@@ -236,6 +236,19 @@ function rt_run_test(RS_type::AbstractRamanType,
     rt_run(RS_type,model,iBand)
 end
 
+# Mockup if no Raman type is chosen:
+function rt_run_ss(model::vSmartMOM_Model; i_band::Integer = 1)
+    rt_run_ss(noRS(), model, i_band)
+end
+
+# Just to make sure we still have it:
+function rt_run_test_ss(RS_type::AbstractRamanType, 
+        model::vSmartMOM_Model, 
+        iBand)
+    rt_run_ss(RS_type,model,iBand)
+end
+
+# Full multiple scattering
 function rt_run(RS_type::AbstractRamanType, 
                     model::vSmartMOM_Model, iBand)
     @unpack obs_alt, sza, vza, vaz = model.obs_geom   # Observational geometry properties
@@ -389,4 +402,160 @@ function rt_run(RS_type::AbstractRamanType,
 
     # Return R_SFI or R, depending on the flag
     return SFI ? (R_SFI, T_SFI, ieR_SFI, ieT_SFI) : (R, T)
+end
+
+# Single scattering only
+function rt_run_ss(RS_type::AbstractRamanType, 
+    model::vSmartMOM_Model, iBand)
+@unpack obs_alt, sza, vza, vaz = model.obs_geom   # Observational geometry properties
+@unpack qp_μ, wt_μ, qp_μN, wt_μN, iμ₀Nstart, μ₀, iμ₀, Nquad = model.quad_points # All quadrature points
+pol_type = model.params.polarization_type
+@unpack max_m = model.params
+@unpack quad_points = model
+
+if obs_alt != 0
+return rt_run_test_ms_ss(RS_type, model, iBand)
+end
+
+# Also to be changed!!
+brdf = model.params.brdf[iBand[1]]
+@unpack ϖ_Cabannes, F₀ = RS_type
+
+
+FT = eltype(sza)                    # Get the float-type to use
+Nz = length(model.profile.p_full)   # Number of vertical slices
+# CFRANKEN NEEDS to be changed for concatenated arrays!!
+
+
+RS_type.bandSpecLim = [] # (1:τ_abs[iB])#zeros(Int64, iBand, 2) #Suniti: how to do this?
+#Suniti: make bandSpecLim a part of RS_type (including noRS) so that it can be passed into rt_kernel and elemental/doubling/interaction and postprocessing_vza without major syntax changes
+#put this code in model_from_parameters
+nSpec = 0;
+for iB in iBand
+nSpec0 = nSpec+1;
+nSpec += size(model.τ_abs[iB], 1); # Number of spectral points
+push!(RS_type.bandSpecLim,nSpec0:nSpec);                
+end
+
+arr_type = array_type(model.params.architecture) # Type of array to use
+SFI = true                          # SFI flag
+NquadN = Nquad * pol_type.n         # Nquad (multiplied by Stokes n)
+dims = (NquadN,NquadN)              # nxn dims
+
+# Need to check this a bit better in the future!
+#FT_dual = length(model.τ_aer[1][1]) > 0 ? typeof(model.τ_aer[1][1]) : FT
+FT_dual = FT
+# Output variables: Reflected and transmitted solar irradiation at TOA and BOA respectively # Might need Dual later!!
+#Suniti: consider adding a new dimension (iBand) to these arrays. The assignment of simulated spectra to their specific bands will take place after batch operations, thereby leaving the computational time unaffected 
+R       = zeros(FT_dual, length(vza), pol_type.n, nSpec)
[…]
