### 2022-09-07 f6824d6
Add contribution statement to README
