### 2022-09-07 fd4eff6
Fix CPU/GPU paramters input
