+++ src/CoreRT/parameters_from_yaml.jl
@@ -94,7 +94,7 @@ function validate_yaml_parameters(params)
               (["radiative_transfer", "l_trunc"], Integer),
               (["radiative_transfer", "depol"], Real),
               (["radiative_transfer", "float_type"], String),
-              (["radiative_transfer", "architecture"], String),
+              (["radiative_transfer", "architecture"], String, ["default_architecture", "Architectures.GPU()", "Architectures.CPU()", "GPU()", "CPU()"]),
 
               # geometry group
               (["geometry", "sza"], Real),
@@ -126,7 +126,8 @@ function validate_yaml_parameters(params)
     for i in 1:length(fields)
         key, elty = fields[i][1:2]
         valid_options = length(fields[i]) == 3 ? fields[i][3] : Array{String}([])
-
+        # @show fields
+        # return 
         # Check if the section exists in the YAML file and is an optional section
         if ("absorption" in keys(params) && key[1] == "absorption") || 
            ("scattering" in keys(params) && key[1] == "scattering") || 
@@ -160,7 +161,11 @@ function parameters_from_yaml(file_path)
     quadrature_type = eval(Meta.parse(params_dict["radiative_transfer"]["quadrature_type"]))
     polarization_type = eval(Meta.parse(params_dict["radiative_transfer"]["polarization_type"]))
     FT = eval(Meta.parse(params_dict["radiative_transfer"]["float_type"]))
-    architecture = eval(Meta.parse(params_dict["radiative_transfer"]["architecture"]))
+
+    arch_string = params_dict["radiative_transfer"]["architecture"]
+    arch_string = !startswith(arch_string, "Architectures.") ? "Architectures." * arch_string : arch_string
+    
+    architecture = eval(Meta.parse(arch_string))
 
     # atmospheric_profile group
     T = convert.(FT, params_dict["atmospheric_profile"]["T"]) # Level
