+++ demo/demo_08_21.md
@@ -1,111 +1,0 @@
-<!-- omit in toc -->
-# VS Code & CrossSection Module Demo
-**Friday, August 21st, 2020**
-
-Hello! 
-
-The goal of this document is to provide an in-text version of all the things I am showing off at today's lab meeting. Namely, I would like to document the features I showcase in **VSCode** and in the Julia **CrossSection** module I created as part of the RadiativeTransfer package I am working on. 
-
-Let's get started! 
-
-<!-- omit in toc -->
-## Table of Contents
-
-- [VSCode](#vscode)
-  - [Introduction](#introduction)
-  - [Getting Started](#getting-started)
-  - [Features List](#features-list)
-- [Absorption](#absorption)
-  - [Description](#description)
-  - [Demo](#demo)
-
-## VSCode 
-
-### Introduction
-
-When I first think of Visual Studio (the "VS" in VSCode), I immediately think of the clunky Microsoft IDE that I used many years ago to write C++ code. It was quite difficult to use and frankly, neither elegant nor approachable. 
-
-In this document, I advocate for **Visual Studio Code**, which is confusingly, much different from **Visual Studio**. The former is a lightweight code editor like Sublime Text or Atom, whereas the latter is a full IDE like Eclipse or NetBeans. 
-
-The primary reason I am vouching for VSCode, the code editor, is that it has completely improved my remote work experience. Previously, I was using `scp` or `git` pushing/pulling regularly to keep files in sync between my local system and remote servers (Fluo/Tofu). 
-
-Now using VSCode, I can directly edit files on-server using a graphical code editor like Sublime. It allows you to perform large computations on Tofu and then directly view resulting plots inside the editor. The code editor is also language-agnostic, so there is great support for all mainstream coding languages. You can even run a Jupyter notebook inside VSCode. I can go on and on... Please see the features list [below](#features-list). 
-
-### Getting Started
-
-1. **Required:** You can download VSCode [here](https://code.visualstudio.com).
-2. **Recommended:** Details on setting up the Remote SSH Extension to VSCode can be found [here](https://code.visualstudio.com/docs/remote/ssh). I would *highly* recommend this step to everyone! It has made working from home so much more pleasant for me. 
-3. **Recommended:** You can set up language-specific extensions via the VSCode Extensions Marketplace (everything is free!). Linking the [Julia](https://marketplace.visualstudio.com/items?itemName=julialang.language-julia), [Python](https://marketplace.visualstudio.com/items?itemName=ms-python.python), [R](https://marketplace.visualstudio.com/items?itemName=Ikuyadeu.r), [Matlab](https://marketplace.visualstudio.com/items?itemName=Gimly81.matlab), and [other](https://marketplace.visualstudio.com/search?target=VSCode&category=Programming%20Languages&sortBy=Installs) languages.  
-4. Optional: You can set up Jupyter Notebooks for Julia/Python/R inside VSCode. Although you can use any of these languages inside Jupyter after it is installed, Jupyter itself is installed only through the Python language extension found [here](https://marketplace.visualstudio.com/items?itemName=ms-python.python). After it is set up, you can switch the kernel/language in the top right corner of an open notebook. Fun fact: It's named Jupyter after the three languages it was built to support: JUlia, PYThon and R. 
-5. Optional: There is a cool VSCode extension called [Live Share](https://marketplace.visualstudio.com/items?itemName=MS-vsliveshare.vsliveshare-pack) that allows multiple users to collaboratively edit code and debug in real-time (sort of like Google Docs but for editing and running code). I have not used this, but I have heard great things from the other Schmidt Scholars that this works really well. 
-
-### Features List
-
[…]
