### 2022-09-07 e87eba2
Add Canopy input from JSON
