+++ paper.md
@@ -27,19 +27,23 @@ bibliography: paper.bib
 
 # Summary
 
-Remote sensing researchers use radiative transfer modeling to interpret satellite data for studying Earth's atmospheric and surface properties. The field plays a key role in how scientists understand many aspects of our rapidly changing planet – from climate change and pollution to the carbon and water cycles. Astronomers use radiative transfer to study the atmospheres of stars and substellar objects like brown dwarfs and exoplanets. 
+Remote sensing researchers use radiative transfer modeling to interpret satellite data for studying Earth's atmospheric and surface properties. The field plays a key role in how scientists understand many aspects of our rapidly changing planet – from climate change and pollution to the carbon and water cycles. Astronomers use radiative transfer to study the atmospheres of stars and substellar objects such as brown dwarfs and exoplanets. 
 
-**vSmartMOM.jl** is a [Julia](https://julialang.org) package that enables the fast computation of atmospheric optical properties and fully-polarized multiple-scattering radiance simulations [@Sanghavi:2013a], based on the Matrix Operator Method, also known as Discrete Space Theory [@grant1969adiscrete; @grant1969bdiscrete; @hunt1969discrete]. Users are free to customize simulation parameters and atmospheric properties, including trace-gas profiles, aerosol distributions, microphysical properties, surface reflectance, and quadrature schemes. Independent submodules can also be imported individually; for example, **Absorption.jl** can be used for computing gaseous absorption and **Scattering.jl** for computing scattering phase-functions. 
+**vSmartMOM.jl** is a [Julia](https://julialang.org) package that enables the fast computation of atmospheric optical properties and fully polarized multiple-scattering radiance simulations [@Sanghavi:2013a], based on the Matrix Operator Method, also known as Discrete Space Theory [@grant1969adiscrete; @grant1969bdiscrete; @hunt1969discrete]. Users are free to customize simulation parameters and atmospheric properties, including trace-gas profiles, aerosol distributions, microphysical properties, surface reflectance, and quadrature schemes. Independent submodules can also be imported individually; for example, **Absorption.jl** can be used for computing gaseous absorption and **Scattering.jl** for computing scattering phase-functions. 
 
-The Julia language provides many exciting opportunities to modernize radiative transfer software. Using the ForwardDiff.jl package [@Revels:2016], Jacobians can be calculated alongside computations using automatic differentiation, allowing for elegant and straightforward parameter-fitting. Julia's multiple dispatch paradigm enables the software architecture to be clean, flexible and reusable. Additionally, optimized techniques have been implemented to speed up the package’s performance on both CPU and GPU by orders-of-magnitude compared to existing radiative transfer codes. 
+The Julia language provides many exciting opportunities to modernize radiative transfer software. For example, using the ForwardDiff.jl package [@Revels:2016], Jacobians can be calculated alongside computations using a technique called automatic differentiation. This technique uses the chain rule, dual numbers, and Julia's multiple dispatch paradigm to propagate not only the numerical computations at each step, but also the derivative. 
 
-**vSmartMOM.jl** has already been used in research projects, ranging from methane-plume simulation to aerosol profile fitting and spectropolarimetric simulations of brown dwarfs and exoplanets. It has also been used in graduate-level remote sensing coursework. Ultimately, **vSmartMOM.jl** aims to accelerate the pace of atmospheric research through efficient software while lowering the barrier-of-entry for researchers and students in remote sensing. 
+We further use Julia's multiple dispatch feature to create a software architecture that is clean, flexible, and reusable. 
+
[…]
