### 2022-09-09 73123bd
Fixminor joss suggestions in docs
