+++ docs/src/pages/Scattering/Overview.md
@@ -19,4 +19,4 @@ For a full demo of how to use this module, please see the [example](https://remo
 
 The Scattering.jl architecture closely follows the user's workflow to calculate the scattering phase-function. There are functions for creating an aerosol, defining scattering parameters, calculating aerosol optical properties, and constructing the phase-function from said optical properties. 
 
-The aerosol optical properties contain computed "Greek Coefficients", which are to be multiplied by matrices composed of generalized spherical functions, in order to produce the phase-functions. Since calculating the Greek Coefficients is the most computationally-intensive part – and the output phase-function can be produced at various resolutions – the [`reconstruct_phase`](@ref) function is separate from [`compute_aerosol_optical_properties`](@ref). 
+The aerosol optical properties contain computed "Greek Coefficients", which are to be multiplied by matrices composed of generalized spherical functions, in order to produce the phase-functions. Since calculating the Greek Coefficients is the most computationally intensive part – and the output phase-function can be produced at various resolutions – the [`reconstruct_phase`](@ref) function is separate from [`compute_aerosol_optical_properties`](@ref). 
