### 2022-09-09 b694c5e
Fixminor joss suggestions in docs
