+++ src/CoreRT/DefaultParameters.yaml
@@ -87,7 +87,7 @@ absorption:
 scattering:
   # List of scattering aerosols and their properties
   aerosols:       
-    - τ_ref:          1.0           # Reference τ
+    - τ_ref:          1.0           # Reference τ (µm)
       μ:              1.3           # Log mean radius (µm)
       σ:              2.0           # Log stddev of radius (µm)
       nᵣ:             1.3           # Real part of refractive index
