### 2022-09-14 d6a19d7
Minor documentation fixes for joss
