### 2022-09-20 1e154b9
Benchmark  for I/I0 with lambertian reflectance
