+++ src/Inelastic/raman_atmo_prop.jl
@@ -178,6 +178,7 @@ function getRamanSSProp!(
         λ = nm_per_m/(0.5*(_grid_in[1]+_grid_in[end]))
         
         if iB==1
+            @show InelasticScattering.compute_ϖ_Cabannes(RS_type, λ_inc, n2, o2)
             ϖ_Cabannes[iB] = InelasticScattering.compute_ϖ_Cabannes(RS_type, λ_inc, n2, o2)
             #@show ϖ_Cabannes
         else
@@ -398,5 +399,157 @@ i_λ₁λ₀_all = unique(cat(i_λ₁λ₀, i_λ₁λ₀_VS_n2, i_λ₁λ₀_VS_
 return nothing
 end
 
+function getRamanSSProp!(
+    RS_type::Union{VS_0to1_plus, VS_1to0_plus}, λ_inc,  target_grid)
 
+    @unpack n2,o2,
+        iBand, grid_in, bandSpecLim, 
+        greek_raman, greek_raman_VS_n2, greek_raman_VS_o2,
+        ϖ_Cabannes, fscattRayl,
+        ϖ_λ₁λ₀, i_λ₁λ₀,
+        ϖ_λ₁λ₀_VS_n2, i_λ₁λ₀_VS_n2,
+        ϖ_λ₁λ₀_VS_o2, i_λ₁λ₀_VS_o2,
+        i_λ₁λ₀_all =  RS_type
+    #n2, o2 = getRamanAtmoConstants(1.e7/λ, T)
+    iBand = []
+    grid_in = []
+    bandSpecLim = []
+
+    nm_per_m = 1.e7;
+    greek_raman = get_greek_raman(RS_type, n2, o2)
+    greek_raman_VS_n2 = get_greek_raman_VS(RS_type, n2)
+    greek_raman_VS_o2 = get_greek_raman_VS(RS_type, o2)
+    #λ_scatt = 2.e7/(grid_in[1]+grid_in[end])     
+    # determine Rayleigh scattering cross-section at incident wavelength λ 
+    atmo_σ_Rayl = compute_optical_Rayl(λ_inc, n2, o2)
+    # determine VRS_0to1 bands for λ₀ due to n2 and o2  
+    nBand = 2; #for n2 and o2 each
+    iBand = [1,2];
+    ν̄ = nm_per_m/λ_inc
+    push!(grid_in, ν̄:0.3:ν̄)
+    push!(grid_in, target_grid)
+    molec = [n2,o2]
+    #========================================================================#
+    #=
+    for mol in molec
+
+        y1 = mol.effCoeff.Δν̃_RoVibRaman_coeff_0to1_JtoJm2;
+        y2 = mol.effCoeff.Δν̃_RoVibRaman_coeff_0to1_JtoJp2;
+
+        band_min = minimum([minimum(y1[y1.!=0]), minimum(y2[y2.!=0])]);
+        band_min += nm_per_m/λ_inc;
+        band_min -= 2.;
+
+        band_max = maximum([maximum(y1[y1.!=0]), maximum(y2[y2.!=0])]);
+        band_max += nm_per_m/λ_inc;
+        band_max += 2.;
+
+        push!(grid_in, band_min:0.3:band_max)
+
+    end
+    =#
+    #========================================================================#
+    FT = eltype(λ_inc);
+    t_w_VS_n2 = Vector{Vector{FT}}(undef,0)
+    t_i_VS_n2 = Vector{Vector{Int}}(undef,0)
+    t_w_VS_o2 = Vector{Vector{FT}}(undef,0)
[…]
