+++ test/benchmarks/prototype_VS_O2Aband.jl
@@ -1,0 +1,327 @@
+##
+using Revise
+using vSmartMOM, vSmartMOM.CoreRT, vSmartMOM.SolarModel
+using vSmartMOM.InelasticScattering
+using Statistics
+using Interpolations
+using InstrumentOperator #for convolution of hires spectrum to instrument grid
+using ImageFiltering
+using Distributions
+using Plots
+
+Fraunhofer=true
+Tsolar = solar_transmission_from_file("/home/sanghavi/code/github/vSmartMOM.jl/src/SolarModel/solar.out")
+Tsolar_interp = LinearInterpolation(Tsolar[4:end, 1], Tsolar[4:end, 2])
+        
+ν₀ = (1e7/730.):0.5:(1e7/630.) #Spectral range of sources of VS  
+#ν₀ = (1e7/750.):0.5:(1e7/731.5) #Spectral range of sources of VS  
+λ₀ = 2e7/(ν₀[1]+ν₀[end]) #400. #nm
+ii₀ = findmin(abs.(1e7./ν₀ .- λ₀))[2]
+νₐ = (1e7/780.):0.5:(1e7/750.) #Target spectral range (O2 A-band)  
+# Computing Fraunhofer Spectrum 
+T_sun = 5777. # KRS 
+P₀ = planck_spectrum_wn(T_sun, collect(ν₀)) 
+Pₐ = planck_spectrum_wn(T_sun, collect(νₐ))    
+#F₀ = zeros(length(ν₀));
+#ν₀ = 1e7/450.:2.:1e7/375.
+R₀ = zeros(length(ν₀)) 
+Rₐ = zeros(length(νₐ)) 
+#T₀ = zeros(length(ν₀)) 
+ieRₐ = zeros(length(νₐ)) 
+#ieT₀ = zeros(length(ν₀))
+R₀_Q = zeros(length(ν₀)) 
+Rₐ_Q = zeros(length(νₐ)) 
+#T₀_Q = zeros(length(ν₀)) 
+ieRₐ_Q = zeros(length(νₐ)) 
+#ieT₀_Q = zeros(length(ν₀))
+#R₀_U = zeros(3,length(ν₀)) 
+#T₀_U = zeros(length(ν₀)) 
+#ieR₀_U = zeros(3,length(ν₀)) 
+#ieT₀_U = zeros(length(ν₀))
+I₀_conv = zeros(length(ν₀)) 
+Iₐ_conv = zeros(length(νₐ)) 
+ieIₐ_conv = zeros(length(νₐ)) 
+Q₀_conv = zeros(length(ν₀)) 
+Qₐ_conv = zeros(length(νₐ)) 
+ieQₐ_conv = zeros(length(νₐ)) 
+
+#=
+for i=1:length(ν₀)
+    #    ##sol_trans = Tsolar_interp(ν₀[i]);
+    #    F₀[i] = 1.0##sol_trans * P[i];
+    #end 
+    λ₀ = 1e7/ν₀[i]        
+    FT = Float64
+    #λ₀ = 1e7/ν₀[ctr]
+    #iBand = 1
+    n2,o2 = InelasticScattering.getRamanAtmoConstants(1.7/λ₀, 300.);
+    RS_type = InelasticScattering.VS_0to1_plus{FT}(
+                n2=n2,
+                o2=o2);
+    # Load YAML files into parameter struct
+    parameters = parameters_from_yaml("test/test_parameters/O2ParametersVS.yaml");
+    # Create model struct (precomputes optical properties) from parameters
[…]
