### 2022-09-21 77497d5
Add dolp plotting and I/I0 output for multisensor geometries
