+++ test/benchmarks/prototype_inelastic_ms.jl
@@ -173,7 +173,7 @@ q2 = plot!(1e7./ν, (Q_conv.+ieQ_conv).*convfct, linewidth = 3, linecolor=:red)
 q3 = plot(1e7./ν, (R[1,2,:].-RnoRS[1,2,:].+ieR[1,2,:]).*convfct, linecolor=:black, xlabel = "λ [nm]")
 q3 = plot!(1e7./ν, (Q_conv.-Q_conv_noRS.+ieQ_conv).*convfct, linewidth = 3, linecolor=:red)
 plot(p1, q1, p2, q2, p3, q3, layout = l, legend = false, title = ["I₀ (no RS)" "Q₀ (no RS)" "I₁ (with RS)" "Q₁ (with RS)" "I₁-I₀" "Q₁-Q₀"], titlefont = font(10))
-savefig("RingEffect63U.png")
+#savefig("RingEffect63U.png")
 
 l = @layout [a1 a2]
 p1 = plot(1e7./ν, 100*ieR[1,1,:]./R[1,1,:].*convfct, linecolor=:black)
@@ -181,4 +181,119 @@ p1 = plot!(1e7./ν, 100*ieI_conv./I_conv.*convfct, linewidth = 2, linecolor=:red
 q1 = plot(1e7./ν, 100*ieR[1,2,:]./R[1,2,:].*convfct, linecolor=:black)
 q1 = plot!(1e7./ν, 100*ieQ_conv./Q_conv.*convfct, linewidth = 2, linecolor=:red, xlabel = "λ [nm]")
 plot(p1, q1, layout = l, legend = false, title = ["Iᵢ/Iₑ [%]" "Qᵢ/Qₑ [%]"], titlefont = font(10))
-savefig("RingSpectrum63U.png")
+#savefig("RingSpectrum63U.png")
+
+using DelimitedFiles
+raylout_rrs = readdlm("out_ray_rrs_385nm_405nm.dat")
+raylout_nors = readdlm("out_ray_nors_385nm_405nm.dat")
+raylout_rrs_ss = readdlm("out_ray_rrs_ss_385nm_405nm.dat")
+aerout_rrs = readdlm("out_aer_rrs_385nm_405nm.dat")
+aerout_nors = readdlm("out_aer_nors_385nm_405nm.dat")
+aerout_rrs_ss = readdlm("out_aer_rrs_ss_385nm_405nm.dat")
+
+x = -40:0.6:40
+#kernel = InstrumentOperator.create_instrument_kernel(Normal(0, 12.5), x) #defining a Gaussian kernel for convulution in wavenumber space
+kernel = InstrumentOperator.create_instrument_kernel(Normal(0, 12.5), x)
+#I_conv = InstrumentOperator.conv_spectra(kernel, )
+
+ν = raylout_nors[:,1]
+#Rayleigh
+rI_noRS = zeros(length(ν))
+rQ_noRS = zeros(length(ν))
+rI = zeros(length(ν))
+rQ = zeros(length(ν))
+ierI = zeros(length(ν))
+ierQ = zeros(length(ν))
+rI_ss = zeros(length(ν))
+rQ_ss = zeros(length(ν))
+ierI_ss = zeros(length(ν))
+ierQ_ss = zeros(length(ν))
+
+rI_conv_noRS = zeros(length(ν))
+rQ_conv_noRS = zeros(length(ν))
+rI_conv = zeros(length(ν))
+rQ_conv = zeros(length(ν))
+ierI_conv = zeros(length(ν))
[…]
