### 2022-09-21 145f1ba
Make minor changes to docs
