### 2022-09-21 e9a0c64
Make minor modifications to paper
