+++ paper.bib
@@ -110,7 +110,7 @@ author = {Hu, Yongxiang and Wielicki, Bruce and Lin, Bing and Gibson, G. and Tsa
 year = {2000},
 month = {05},
 pages = {681-690},
-title = {??-Fit: A fast and accurate treatment of particle scattering phase functions with weighted singular-value decomposition least-squares fitting},
+title = {𝛅-Fit: A fast and accurate treatment of particle scattering phase functions with weighted singular-value decomposition least-squares fitting},
 volume = {65},
 journal = {Journal of Quantitative Spectroscopy & Radiative Transfer - J QUANT SPECTROSC RADIAT},
 doi = {10.1016/S0022-4073(99)00147-8}
@@ -249,4 +249,44 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   pages={2139--2150},
   year={2010},
   publisher={Elsevier}
+}
+
+@article{goody:1989,
+title = {The correlated-k method for radiation calculations in nonhomogeneous atmospheres},
+journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+volume = {42},
+number = {6},
+pages = {539-550},
+year = {1989},
+issn = {0022-4073},
+doi = {https://doi.org/10.1016/0022-4073(89)90044-7},
+url = {https://www.sciencedirect.com/science/article/pii/0022407389900447},
+author = {Richard Goody and Robert West and Luke Chen and David Crisp},
+abstract = {We discuss the accuracy of the correlated-k method for radiation calculations with spectrally averaged data in nonhomogeneous atmospheres. Correlated-k may be justified in special cases but, otherwise, numerical comparisons must be made with exact calcualtions. Comparisons for scattering and absorbing atmospheres containing CO2, H2O and O3 lead to the following conclusions: (i) the errors in correlated-k are generally of order of magnitude 1%. (ii) Much larger errors occur only when a radiative quantity is very much smaller than its average value. (iii) Errors do not depend systematically on the gas molecule, the distributions of gases and aerosols, or on the aerosol optical properties. (iv) Errors do not systematically increase with the order of differencing. We show that the multiplication property for the transmission by overlapping bands can be incorporated into correlated-k; that temperature effects can be interpolated on a coarse grid; and that 10 quadrature points are often sufficient to average over complex spectral intervals containing thousands of lines. Correlated-k is well suited to iterative methods involved in the inversion of satellite radiances or in numerical weather or climate models.}
+}
+
+@article{liu:2016,
+  author={Liu, X. and Yang, Q. and Li, H. and Jin, Z. and Wu, W. and Kizer, S. and Zhou, D. K. and Yang, P.},
+  title={Development of a fast and accurate PCRTM radiative transfer model in the solar spectral region},
+  year={2016},
+  journal={Appl. Opt.},
+  volume={55},
+  number={29},
+  pages={8236--8247},
+  doi={10.1364/AO.55.008236},
+}
+
+@article{spurr:2011,
+title = {A linearized two-stream radiative transfer code for fast approximation of multiple-scatter fields},
+journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+volume = {112},
+number = {16},
+pages = {2630-2637},
+year = {2011},
+issn = {0022-4073},
+doi = {https://doi.org/10.1016/j.jqsrt.2011.06.014},
+url = {https://www.sciencedirect.com/science/article/pii/S0022407311002263},
+author = {R. Spurr and V. Natraj},
+keywords = {2-Stream, Radiative transfer, Jacobians, Multiple-scatter accuracy},
[…]
