+++ Project.toml
@@ -24,7 +24,6 @@ NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 NNlib = "872c559c-99b0-510c-b3b7-b6c96a88d5cd"
 NaNMath = "77ba4419-2d1f-58cd-9bb1-8ffee604a2e3"
 NetCDF = "30363a11-5582-574a-97bb-aa9a979735b9"
-NetCDF_jll = "7243133f-43d8-5620-bbf4-c2c921802cf3"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
@@ -55,10 +54,7 @@ Interpolations = "0.12, 0.13"
 JLD2 = "0.1, 0.2, 0.3, 0.4"
 JSON = "0.21"
 KernelAbstractions = "0.6, 0.7, 0.8"
-NCDatasets = "0.11, 0.12"
 NNlib = "0.8"
-NetCDF = "0.10, 0.11"
-NetCDF_jll = "400.701.400 - 400.702.400"
 Parameters = "0.12"
 Polynomials = "2, 3"
 ProgressMeter = "1.3"
