+++ test/benchmarks/prototype_inelastic_ms.jl
@@ -1,11 +1,20 @@
 ##
 using Revise
-using vSmartMOM, vSmartMOM.CoreRT
+using vSmartMOM, vSmartMOM.CoreRT, vSmartMOM.SolarModel
 using vSmartMOM.InelasticScattering
 using Statistics
+using Interpolations
+using InstrumentOperator #for convolution of hires spectrum to instrument grid
+using ImageFiltering
+using Distributions
+using Plots
+# Benchmarks: http://web.gps.caltech.edu/~vijay/Rayleigh_Scattering_Tables/STOKES/
+##
 
 # Load YAML files into parameter struct
-parameters = parameters_from_yaml("test/test_parameters/O2Parameters2.yaml");
+parameters = 
+    parameters_from_yaml("test/test_parameters/FraunhoferMockParameters.yaml");
+#parameters = parameters_from_yaml("test/test_parameters/O2Parameters2.yaml");
 # Create model struct (precomputes optical properties) from parameters
 model      = model_from_parameters(parameters);
 
@@ -37,7 +46,9 @@ RS_type = InelasticScattering.RRS(
             Z⁻⁺_λ₁λ₀    = zeros(FT,1,1), 
             Z⁺⁺_λ₁λ₀    = zeros(FT,1,1), 
             i_ref       = i_ref,
-            n_Raman=0);
+            n_Raman     = 0,
+            F₀          = zeros(FT,1,1));
+
 
 #vSmartMOM.get_greek_raman!(RS_type, n2, o2);
 # now compute other optical parameters for Raman:
@@ -45,7 +56,19 @@ RS_type = InelasticScattering.RRS(
 #RS_type = vSmartMOM.RRS(...)
 # Compute Raman SSA properties:
 CoreRT.getRamanSSProp!(RS_type, 1e7/ν̃, ν);
+T_sun = 5777. # K
+P = planck_spectrum_wn(T_sun, ν)
+Tsolar = solar_transmission_from_file("/home/sanghavi/code/github/vSmartMOM.jl/src/SolarModel/solar.out")
+Tsolar_interp = LinearInterpolation(Tsolar[4:end, 1], Tsolar[4:end, 2])
+F₀ = zeros(length(P));
+RS_type.F₀ = zeros(model.params.polarization_type.n, length(P))
+for i=1:length(P)
+    sol_trans = Tsolar_interp(ν[i]);
+    F₀[i] = sol_trans * P[i];
+    RS_type.F₀[1,i] = F₀[i];
+end 
 
+#solar_transmission_from_file(file_name::String) = readdlm(file_name)
 # For now, convert these special cases to the right array type:
 #aType = array_type(model.params.architecture);
 #RS_type.Z⁺⁺_λ₁λ₀ = aType(RS_type.Z⁺⁺_λ₁λ₀);
@@ -83,13 +106,32 @@ RnoRS, TnoRS, _, _ = CoreRT.rt_run_ms(noRS(),
             model.τ_abs[1],
             model.params.brdf[1],
             model.params.arhitecture);
-=#            
+   =#       
[…]
