+++ src/CoreRT/CoreKernel/elemental_canopy.jl
@@ -1,0 +1,153 @@
+#=
+ 
+This file contains RT elemental-related functions
+ 
+=#
+"Elemental single-scattering layer"
+function elemental!(pol_type, SFI::Bool, 
+                            τ_sum::AbstractArray,#{FT2,1}, #Suniti
+                            dτ::AbstractArray,
+                            computed_layer_properties::CoreDirectionalScatteringOpticalProperties,
+                            m::Int,                     # m: fourier moment
+                            ndoubl::Int,                # ndoubl: number of doubling computations needed 
+                            scatter::Bool,              # scatter: flag indicating scattering
+                            quad_points::QuadPoints{FT2}, # struct with quadrature points, weights, 
+                            added_layer::Union{AddedLayer{FT},AddedLayerRS{FT}}, 
+                            architecture) where {FT<:Union{AbstractFloat, ForwardDiff.Dual},FT2,M}
+
+    @unpack r⁺⁻, r⁻⁺, t⁻⁻, t⁺⁺, j₀⁺, j₀⁻ = added_layer
+    @unpack qp_μ, iμ₀, wt_μN, qp_μN = quad_points
+    @unpack τ, ϖ, Z⁺⁺, Z⁻⁺, G = computed_layer_properties
+    
+    arr_type = array_type(architecture)
+
+    # Need to check with paper nomenclature. This is basically eqs. 19-20 in vSmartMOM
+    I₀    = arr_type(pol_type.I₀)
+    D     = Diagonal(arr_type(repeat(pol_type.D, size(qp_μ,1))))
+
+    device = devi(architecture)
+
+    # If in scattering mode:
+    if scatter
+        # for m==0, ₀∫²ᵖⁱ cos²(mϕ)dϕ/4π = 0.5, while
+        # for m>0,  ₀∫²ᵖⁱ cos²(mϕ)dϕ/4π = 0.25  
+        wct02 = m == 0 ? FT(0.50)              : FT(0.25)
+        wct2  = m == 0 ? wt_μN/2               : wt_μN/4
+ 
+        # More computationally intensive definition of a single scattering layer with variable (0-∞) absorption
+        # with absorption in batch mode, low tau_scatt but higher tau_total, needs exact equations
+        kernel! = get_canopy_elem_rt!(device)
+        event = kernel!(r⁻⁺, t⁺⁺, ϖ, dτ, G, Z⁻⁺, Z⁺⁺, qp_μN, wct2, ndrange=size(r⁻⁺)); 
+        wait(device, event)
+        synchronize_if_gpu()
+
+        # SFI part
+        kernel! = get_canopy_elem_rt_SFI!(device)
[…]
