### 2022-09-25 e006281
Added canopy elemental function
