### 2022-10-04 ad970d0
Add more info to absirption tutorial
