+++ docs/src/pages/tutorials/Tutorial_Absorption.jl
@@ -1,19 +1,17 @@
 # # Absorption: Spectral Line Shapes
 
-# Part of ESE 156 Remote Sensing Class; 2020; Christian Frankenberg
-
-# The purpose of this tutorial is to demonstrate how to compute line-shapes (in our case mostly of rotational-vibrational transition lines), what processes determine line-shapes, and how the dependencies in the line-shape on pressure and temperature could actually be used to our advantage in a retrieval system.
+# The purpose of this tutorial is to demonstrate how to compute line-shapes using the vSmartMOM Julia package. Here, we focus on rotational-vibrational transition lines, what processes determine line-shapes, and how the line-shapes depend on pressure and temperature.
 
 # ### Basic Tools
 
 # We are making use of the [HITRAN database](https://www.cfa.harvard.edu/hitran/), see a list of tutorials [here](https://www.youtube.com/playlist?list=PLqOG3cBizTUFQw5lwpkCpQFCVuxHTZgrT).
 
-# There are other spectroscopic linelists, but we are focusing on HITRAN only here. In addition, there are more complex line-shapes that we are not treating here, including [collisional narrowing](https://journals.aps.org/pr/abstract/10.1103/PhysRev.89.472) and [line-mixing](https://aip.scitation.org/doi/abs/10.1063/1.478724), [collision-induced absorption (CIA)](https://www.sciencedirect.com/science/article/abs/pii/S0022407311003773).
+# There are other spectroscopic linelists, but we are focusing on HITRAN only here. In addition, there are more complex line-shapes that we are not treating here, including [collisional narrowing](https://journals.aps.org/pr/abstract/10.1103/PhysRev.89.472) and [line-mixing](https://aip.scitation.org/doi/abs/10.1063/1.478724), [collision-induced absorption (CIA)](https://www.sciencedirect.com/science/article/abs/pii/S0022407311003773). In the future, these might be added to vSmartMOM.
 
-# ### Line-shape (background covered in class):
+# ### Line-shape:
 # If we have a line-strength S (in cm$^{-1}$ cm$^2$/molecule) for a specific transition at $\nu_0$, we can compute the cross section as:
 
-# $$\sigma(\nu) = S\Phi(\nu-\nu_0)\,,$$
+# $$\sigma(\nu) = S \cdot \Phi(\nu-\nu_0)\,,$$
 
 # where $\Phi(\nu-\nu_0)$ denotes the line-shape function (in 1/cm$^{-1}$), which is normalized to 1: 
 
@@ -52,25 +50,27 @@ println("Doppler shift = $(1e7/(v₀-Δ_ν)-1e7/v₀) nm")
 
 # #### Random movements of molecules in gases lead to doppler broadening effects
 
-# In the one-dimensional case along say the x-axis (we only observe the distribution along one line of sight), the speed of molecules is distributed according to the Maxwell-Boltzmann distribution:
+# In the one-dimensional case -- say along the x-axis --, the speed of molecules is distributed according to the Maxwell-Boltzmann distribution:
 
-# $$f(v_x) = \sqrt{\frac{m}{2\pi kT}}\exp{\left(-\frac{mv_x^2}{2kT}\right)}.$$
+# $$f(v_x) = \sqrt{\frac{m}{2\pi kT}}\exp{\left(-\frac{mv_x^2}{2kT}\right)}\,,$$
 
-# We can then define a Doppler width $\Delta \nu_D$ as
+# where $m$ is the particle mass, $k$ is the Boltzmann constant, and T is thermodynamic temperature.
+
+# We can thus define a Doppler width $\Delta \nu_D$ as
 
[…]
