### 2022-10-04 a533c57
Describe plots in Scattering tutorial
