+++ src/CoreRT/postprocessing_vza_ms.jl
@@ -175,3 +175,101 @@ function postprocessing_vza_ms!(RS_type::Union{RRS, VS_0to1_plus, VS_1to0_plus},
         end
     end
 end
+
+"Perform post-processing to azimuthally-weight RT matrices"
+function postprocessing_vza_ms_canopy!(RS_type::noRS, 
+        sensor_levels, 
+        quad_points,
+        iμ₀, pol_type, 
+        composite_layer, 
+        solJ₀,
+        vza, qp_μ, m, vaz, μ₀, weight, 
+        nSpec, SFI, 
+        uwJ, dwJ, uwieJ, dwieJ, 
+        hdr_J₀⁻, bhr_uw, bhr_dw,
+        I_static::AbstractArray{FT2}, 
+        arr_type) where {FT<:Union{AbstractFloat, ForwardDiff.Dual},FT2}
+    
+    # idfx of μ0 = cos(sza)
+    st_iμ0, istart0, iend0 = get_indices(iμ₀, pol_type);
+    
+    # Convert these to Arrays (if CuArrays), so they can be accessed by index
+    topR⁺⁻ = composite_layer.topR⁺⁻;
+    botR⁻⁺ = composite_layer.botR⁻⁺;
+    topJ₀⁺ = composite_layer.topJ₀⁺;
+    topJ₀⁻ = composite_layer.topJ₀⁻;
+    botJ₀⁺ = composite_layer.botJ₀⁺;
+    botJ₀⁻ = composite_layer.botJ₀⁻;
+    
+    tuwJ = [zeros(typeof(topJ₀⁺[1][1,1,1]), (length(topJ₀⁺[1][:,1,1]), 1, nSpec)) for i=1:length(sensor_levels)] #similar(topJ₀⁺); #deepcopy(topJ₀⁺)
+    tdwJ = [zeros(typeof(topJ₀⁺[1][1,1,1]), (length(topJ₀⁺[1][:,1,1]), 1, nSpec)) for i=1:length(sensor_levels)]#similar(topJ₀⁺); #deepcopy(topJ₀⁺)
+    #@show size(tuwJ[1])
+    for ims = 1:length(sensor_levels)
+        if(sensor_levels[ims]==0)
+            tuwJ[ims] .= botJ₀⁻[ims]
+            tdwJ[ims] .= botJ₀⁺[ims]
+            #if(m==2)
+            #    @show botJ₀⁺[ims][:,1,end]
+            #    @show tdwJ[ims][:,1,end]
+            #    showp
+            #end
+            
+        else
+            #@show size(topR⁺⁻[ims]), size(botR⁻⁺[ims]),size(topJ₀⁺[ims]),size(botJ₀⁻[ims])#,size(tdwJ[ims]),size(tuwJ[ims])
+            compute_interlayer_flux!(RS_type,
+                                    I_static, 
+                                    topR⁺⁻[ims], botR⁻⁺[ims], 
+                                    topJ₀⁺[ims], botJ₀⁻[ims],
+                                    tdwJ[ims], tuwJ[ims], arr_type);
+
+            if(m==0)
+                @show topJ₀⁺[ims][:,1,end]
+                @show tdwJ[ims][:,1,end]
+                @show "========="
+                @show botJ₀⁻[ims][:,1,end]
+                @show tuwJ[ims][:,1,end]
+                #showp
+            end
+                                    
+            #tmpR = @timeit "postprocessing_vza inv1" batch_inv!(
+            #        tmp_inv, I_static .- topR⁺⁻[ims,:] ⊠ botR⁻⁺[ims,:]) #Suniti
[…]
