+++ src/CoreRT/rt_run_multisensor.jl
@@ -322,7 +322,9 @@ function rt_run_test_ms(RS_type::AbstractRamanType,
                     ims, RS_type, 
                     scattering_interfaces_all[end], 
                     SFI, 
-                    composite_layer, added_layer_surface, I_static, arr_type)
+                    composite_layer, 
+                    added_layer_surface, 
+                    I_static, arr_type)
             #end
         end
         # @timeit "interaction" interaction_ms!(RS_type,
