+++ src/CoreRT/CoreKernel/rt_kernel.jl
@@ -5,7 +5,16 @@ This file implements rt_kernel!, which performs the core RT routines (elemental,
 =#
 #No Raman (default)
 # Perform the Core RT routines (elemental, doubling, interaction)
-function rt_kernel!(RS_type::noRS, pol_type, SFI, added_layer, composite_layer, computed_layer_properties, m, quad_points, I_static, architecture, qp_μN, iz) 
+function rt_kernel!(RS_type::noRS, 
+            pol_type, SFI, 
+            added_layer, 
+            composite_layer, 
+            computed_layer_properties, 
+            m, 
+            quad_points, 
+            I_static, 
+            architecture, 
+            qp_μN, iz) 
 
     @unpack τ_λ, ϖ_λ, τ, ϖ, Z⁺⁺, Z⁻⁺, dτ_max, dτ, ndoubl, dτ_λ, expk, scatter, τ_sum, scattering_interface = computed_layer_properties
     @show τ, ϖ, dτ_max, ndoubl
@@ -156,18 +165,12 @@ function rt_kernel!(RS_type::Union{RRS, VS_0to1, VS_1to0}, pol_type, SFI, added_
     end
 end
 
-
-
-
-
-
-
-### New update:
+### 
 function rt_kernel!(RS_type::noRS{FT}, 
                     pol_type, SFI, 
                     added_layer, 
                     composite_layer, 
-                    computed_layer_properties::M,# CoreScatteringOpticalProperties, 
+                    computed_layer_properties::M, 
                     scattering_interface, 
                     τ_sum, 
                     m, quad_points, 
@@ -178,33 +181,26 @@ function rt_kernel!(RS_type::noRS{FT},
     @unpack qp_μ, μ₀ = quad_points
     # Just unpack core optical properties from 
     @unpack τ, ϖ, Z⁺⁺, Z⁻⁺ = computed_layer_properties
-    # SUNITI, check? Also, better to write function here
-    #@show τ, ϖ
-    dτ_max = minimum([maximum(τ .* ϖ), FT(0.001) * minimum(qp_μ)])
-
-    _, ndoubl = doubling_number(dτ_max, maximum(τ .* ϖ))
+    
     # @show ndoubl
     scatter = true # edit later
-    arr_type = array_type(architecture)
-    # Compute dτ vector
-    dτ = τ ./ 2^ndoubl
-    expk = arr_type(exp.(-dτ /μ₀))
     
+    dτ, ndoubl, expk = init_layer(computed_layer_properties, quad_points, pol_type, architecture)
 
     # If there is scattering, perform the elemental and doubling steps
     if scatter
         #@show typeof(computed_layer_properties)
         @timeit "elemental" elemental!(pol_type, SFI, 
-                                        τ_sum, dτ, 
-                                        computed_layer_properties, 
-                                        m, ndoubl, scatter, quad_points,  
-                                        added_layer,  architecture)
+                                τ_sum, dτ, 
+                                computed_layer_properties, 
+                                m, ndoubl, scatter, quad_points,  
+                                added_layer,  architecture)
         #println("Elemental done...")
         @timeit "doubling"   doubling!(pol_type, SFI, 
-                                        expk, ndoubl, 
-                                        added_layer,
-                                        I_static, architecture)
-        #println("Doubling done...")
[…]

[…]
