### 2022-10-11 #4575 @RupeshJey
Hi Pierre, just wrapped up on our side!

We’ve addressed the concerns listed in the above reviews. Here’s a mostly-exhaustive list of the changes:

- The CPU/GPU selection bug in the yaml file is fixed. The user can input either Architectures.GPU/CPU or just GPU/CPU now.
- We discuss “why” GPU and auto-differentiation are so useful for our type of radiative transfer simulation.
- There is now a “How to Contribute” section that describes how contributors can use git’s features to contribute.
- Docstrings are fixed in the Methods & Types section.
- Additional references are added to the paper to discuss the codes available before vSmartMOM.
- The technical citation issues have been addressed
- The paper feedback has all been taken into account. The minor revisions have been made. Some new definitions are added (e.g. auto-differentiation). We briefly discuss other codes and differences – “Different codes use various approximations like…”
- Documentation-wise, the links are fixed, the minor suggestions are addressed, and the tutorials’ plots, variables, and topics are better-explained (e.g. Doppler Broadening).
- The demos folder was indeed outdated and has now been removed, and the use-cases are now covered exclusively in the tutorials section of the documentation.

There were a couple of points on the paper that we wanted to note as well:

- The Q and U benchmarks are not stated as percentages because they are so minuscule that small discrepancies result in huge fractional variation. For example, -0.0019 and -0.0020 are quite acceptably close, but differ by 5% error. Not sure how to better address this.
- We tried to create a few charts to show the benchmarks as compared to established literature, but the graphs became quickly confusing with so many quantities to compare. We feel the tables and summary of the accuracy is still the better way to go.

Main branch: https://github.com/RemoteSensingTools/vSmartMOM.jl
Paper md (on joss branch): https://github.com/RemoteSensingTools/vSmartMOM.jl/blob/joss/paper.md
Web documentation: https://remotesensingtools.github.io/vSmartMOM.jl/dev/

Thanks all!
