+++ test/test_CoreRT.jl
@@ -19,7 +19,7 @@
                 parameters.brdf = [vSmartMOM.CoreRT.LambertianSurfaceScalar(ρ)]
                 model = model_from_parameters(parameters);
                 model.τ_rayl[1] .= τ
-                R_modeled[sza_i, az_i, :] = CoreRT.rt_run(model, i_band=1)[1][:,1,1] / model.quad_points.μ₀
+                R_modeled[sza_i, az_i, :] = vSmartMOM.CoreRT.rt_run(model, i_band=1)[1][:,1,1] / model.quad_points.μ₀
                 R_deltas[sza_i, az_i, :] = abs.(R_true[sza_i][az_i] - R_modeled[sza_i, az_i, :]) ./ R_true[sza_i][az_i]
             end
         end
