+++ src/CoreRT/rt_run_canopy.jl
@@ -100,8 +100,8 @@ function rt_run_canopy(RS_type::AbstractRamanType,
             constructCoreOpticalProperties(RS_type,iBand,m,model);
         
         
-        @show BiLambMod
-        𝐙⁺⁺, 𝐙⁻⁺ = CanopyOptics.compute_Z_matrices(BiLambMod, Array(qp_μN), LAD, m)    
+        #@show BiLambMod
+        𝐙⁺⁺, 𝐙⁻⁺ = CanopyOptics.compute_Z_matrices_aniso(BiLambMod, Array(qp_μN), LAD, m)    
         # This basically multiplies with G again, needs to be fixed later (or removed from compute_Z_matrices)
         G1 = arr_type(CanopyOptics.G(Array(qp_μN), LAD))
         #@show G1
@@ -129,7 +129,7 @@ function rt_run_canopy(RS_type::AbstractRamanType,
             # Expand all layer optical properties to their full dimension:
             @timeit "OpticalProps" layer_opt = 
                 expandOpticalProperties(layer_opt_props[iz], arr_type)
-
+            @show layer_opt.τ, layer_opt.ϖ, iz
             # Perform Core RT (doubling/elemental/interaction)
             rt_kernel!(RS_type, pol_type, SFI, 
                         #bandSpecLim, 
@@ -207,21 +207,11 @@ function rt_run_canopy(RS_type::AbstractRamanType,
     #return SFI ? (R_SFI, T_SFI, ieR_SFI, ieT_SFI) : (R, T)
     #end
 end
-function get_solJ_canopy(nSpec, 
-                    pol_type, 
-                    in_τ_sum::AbstractArray{FT}, 
-                    μ₀, arr_type) where {FT, FT2, M} #where {FT<:Union{AbstractFloat, ForwardDiff.Dual},FT2,M}
 
-    solJ₀      = zeros(FT, pol_type.n, nSpec)
-    #@show size(arr_type(pol_type.I₀) .* exp.(-in_τ_sum/μ₀))
-    #@show size(solJ₀), size(pol_type.I₀[1]),  size(in_τ_sum)
-    #for ip=1:pol_type.n
-    #    solJ₀[ip,:] = pol_type.I₀[ip] * exp.(-in_τ_sum/μ₀) #Array(added_layer.j₀⁺[iμ₀:iμ₀+pol_type.n-1,1,:]) 
-    solJ₀ .= Array(arr_type(pol_type.I₀) .* exp.(-in_τ_sum/μ₀))' #Array(added_layer.j₀⁺[iμ₀:iμ₀+pol_type.n-1,1,:]) 
-    
-    #end 
-    
-    return solJ₀;
+function get_solJ_canopy(pol_type, 
+                    in_τ_sum::AbstractArray{FT}, 
+                    μ₀, arr_type) where {FT} 
+    solJ₀ = Array(arr_type(pol_type.I₀) .* exp.(-in_τ_sum/μ₀))' 
 end
 
 #For multisensor use (especially for the computation of TOC parameters)
@@ -310,7 +300,8 @@ function rt_run_canopy_ms(RS_type::AbstractRamanType,
 
     #TODO: if RS_type!=noRS, create ϖ_λ₁λ₀, i_λ₁λ₀, fscattRayl, Z⁺⁺_λ₁λ₀, Z⁻⁺_λ₁λ₀ (for input), and ieJ₀⁺, ieJ₀⁻, ieR⁺⁻, ieR⁻⁺, ieT⁻⁻, ieT⁺⁺, ier⁺⁻, ier⁻⁺, iet⁻⁻, iet⁺⁺ (for output)
     #getRamanSSProp(RS_type, λ, grid_in)
-
[…]
