+++ src/CoreRT/rt_run_multisensor.jl
@@ -10,169 +10,6 @@ Sensors are labeled from TOA->BOA, and each share the same set of viewing angles
 Upward looking viewing angles are denoted by -90<VZA<0 and downward viewing angles
 are denoted by 0<VZA<90.
 """
-#=
-function rt_run_ms(RS_type::AbstractRamanType, #Default - no Raman scattering (noRS)
-                pol_type::AbstractPolarizationType,   # Polarization type (IQUV)
-                obs_geom::ObsGeometry,                # Solar Zenith, Viewing Zenith, Viewing Azimuthal 
-                sensor_levels::Vector{Int64},#::ObsGeometry,           # [0] (R for satellite, T for groundbased, default), 
-                                                      # [Nz] (only groundbased),
-                                                      # [0, (1:Nz-1)] (general case)
-                τ_rayl,                               # Rayleigh optical depth 
-                τ_aer,                                # Aerosol optical depth and single-scattering albedo
-                quad_points::QuadPoints,              # Quadrature points and weights
-                max_m,                                # Max Fourier terms
-                aerosol_optics,                       # AerosolOptics (greek_coefs, ω̃, k, fᵗ)
-                greek_rayleigh::GreekCoefs,           # Greek coefficients of Rayleigh Phase Function
-                τ_abs,                                # nSpec x Nz matrix of absorption
-                brdf,                                 # BRDF surface type
-                architecture::AbstractArchitecture)   # Whether to use CPU / GPU
-
-    @unpack obs_alt, sza, vza, vaz = obs_geom   # Observational geometry properties
-    @unpack qp_μ, wt_μ, qp_μN, wt_μN, iμ₀Nstart, μ₀, iμ₀, Nquad = quad_points # All quadrature points
-    @unpack ϖ_Cabannes = RS_type
-    FT = eltype(sza)                    # Get the float-type to use
-    Nz = length(τ_rayl)                 # Number of vertical slices
-    nSpec = size(τ_abs, 1)              # Number of spectral points
-    arr_type = array_type(architecture) # Type of array to use
-    SFI = true                          # SFI flag
-    NquadN = Nquad * pol_type.n         # Nquad (multiplied by Stokes n)
-    dims = (NquadN,NquadN)              # nxn dims
-    nAer  = length(aerosol_optics)      # Number of aerosols
- 
-    # Need to check this a bit better in the future!
-    FT_dual = length(τ_aer) > 0 ? typeof(τ_aer[1]) : FT
-    @show FT_dual
-
-    # Output variables: Reflected and transmitted solar irradiation at TOA and BOA respectively # Might need Dual later!!
-    #R = zeros(FT_dual, length(sensor_levels), length(vza), pol_type.n, nSpec)
-    #T = zeros(FT_dual, length(sensor_levels), length(vza), pol_type.n, nSpec)
-    #R_SFI 
-    uwJ = zeros(FT_dual, length(sensor_levels), length(vza), pol_type.n, nSpec)
-    #T_SFI 
-    dwJ = zeros(FT_dual, length(sensor_levels), length(vza), pol_type.n, nSpec)
-    #ieR_SFI
-    uwieJ = zeros(FT_dual, length(sensor_levels), length(vza), pol_type.n, nSpec)
-    #ieT_SFI
-    dwieJ = zeros(FT_dual, length(sensor_levels), length(vza), pol_type.n, nSpec)
[…]
