+++ src/CoreRT/CoreKernel/rt_kernel.jl
@@ -19,11 +19,13 @@ function rt_kernel!(RS_type::noRS,
     @unpack τ_λ, ϖ_λ, τ, ϖ, Z⁺⁺, Z⁻⁺, dτ_max, dτ, ndoubl, dτ_λ, expk, scatter, τ_sum, scattering_interface = computed_layer_properties
     @show τ, ϖ, dτ_max, ndoubl
     # If there is scattering, perform the elemental and doubling steps
+    @show "Ha"
     if scatter
         
         @timeit "elemental" elemental!(pol_type, SFI, τ_sum, dτ_λ, dτ, ϖ_λ, ϖ, Z⁺⁺, Z⁻⁺, m, ndoubl, scatter, quad_points,  added_layer,  I_static, architecture)
         #println("Elemental done...")
         @timeit "doubling"   doubling!(pol_type, SFI, expk, ndoubl, added_layer, I_static, architecture)
+        @show added_layer.r⁻⁺[1,1,1], added_layer.r⁺⁻[1,1,1],added_layer.t⁺⁺[1,1,1], added_layer.t⁻⁻[:,:,1] 
         #println("Doubling done...")
     else # This might not work yet on GPU!
         # If not, there is no reflectance. Assign r/t appropriately
@@ -62,12 +64,14 @@ function rt_kernel_canopy!(RS_type::noRS, pol_type, SFI, added_layer, composite_
 
     @unpack τ_λ, ϖ_λ, τ, ϖ, Z⁺⁺, Z⁻⁺, dτ_max, dτ, ndoubl, dτ_λ, expk, scatter, τ_sum, scattering_interface = computed_layer_properties
     @show τ, ϖ, dτ_max, ndoubl
+    @show "Ha2"
     # If there is scattering, perform the elemental and doubling steps
     if scatter
         
         @timeit "elemental_canopy" elemental_canopy!(pol_type, SFI, τ_sum, dτ_λ, dτ, ϖ_λ, ϖ, Z⁺⁺, Z⁻⁺, m, ndoubl, scatter, quad_points,  added_layer,  I_static, architecture)
         #println("Elemental done...")
         @timeit "doubling"   doubling!(pol_type, SFI, expk, ndoubl, added_layer, I_static, architecture)
+        @show added_layer.r⁻⁺[1,1,1], added_layer.r⁺⁻[1,1,1],added_layer.t⁺⁺[1,1,1], added_layer.t⁻⁻[:,:,1] 
         #println("Doubling done...")
     else # This might not work yet on GPU!
         # If not, there is no reflectance. Assign r/t appropriately
@@ -178,6 +182,7 @@ function rt_kernel!(RS_type::noRS{FT},
                     architecture, 
                     qp_μN, iz) where {FT,M}
     #@show array_type(architecture)
+    
     @unpack qp_μ, μ₀ = quad_points
     # Just unpack core optical properties from 
     @unpack τ, ϖ, Z⁺⁺, Z⁻⁺ = computed_layer_properties
@@ -200,6 +205,8 @@ function rt_kernel!(RS_type::noRS{FT},
                                 expk, ndoubl, 
                                 added_layer,
                                 I_static, architecture)
+        #@show added_layer.r⁻⁺[1:2,1,1], added_layer.r⁺⁻[1:2,1,1],added_layer.t⁺⁺[1:2,1,1], added_layer.t⁻⁻[1:2,1,1] 
+        #@show dτ, ndoubl, expk
     #println("Doubling done...")
     else # This might not work yet on GPU!
         # If not, there is no reflectance. Assign r/t appropriately
