+++ src/CoreRT/CoreKernel/rt_kernel_multisensor.jl
@@ -1,238 +1,4 @@
-#=
- 
-This file implements rt_kernel!, which performs the core RT routines (elemental, doubling, interaction)
- 
-=#
-#No Raman (default)
-# Perform the Core RT routines (elemental, doubling, interaction)
-#=
-function rt_kernel_multisensor!(RS_type::noRS, 
-                                sensor_levels, 
-                                pol_type, 
-                                SFI, 
-                                added_layer, 
-                                composite_layer, 
-                                computed_layer_properties, 
-                                m, 
-                                quad_points, 
-                                I_static, 
-                                architecture, 
-                                qp_μN, 
-                                iz, 
-                                arr_type) 
 
-    @unpack τ_λ, ϖ_λ, τ, ϖ, Z⁺⁺, Z⁻⁺, dτ_max, dτ, ndoubl, dτ_λ, expk, scatter, τ_sum, scattering_interface = computed_layer_properties
-
-    # If there is scattering, perform the elemental and doubling steps
-    if scatter
-        
-        @timeit "elemental" elemental!(pol_type, SFI, τ_sum, dτ_λ, dτ, ϖ_λ, ϖ, Z⁺⁺, Z⁻⁺, m, ndoubl, scatter, quad_points,  added_layer,  I_static, architecture)
-        #println("Elemental done...")
-        @timeit "doubling"   doubling!(pol_type, SFI, expk, ndoubl, added_layer, I_static, architecture)
-        #println("Doubling done...")
-    else # This might not work yet on GPU!
-        # If not, there is no reflectance. Assign r/t appropriately
-        added_layer.r⁻⁺[:] .= 0;
-        added_layer.r⁺⁻[:] .= 0;
-        added_layer.j₀⁻[:] .= 0;
-        temp = Array(exp.(-τ_λ./qp_μN'))
-        #added_layer.t⁺⁺, added_layer.t⁻⁻ = (Diagonal(exp(-τ_λ / qp_μN)), Diagonal(exp(-τ_λ / qp_μN)))   
-        for iλ = 1:length(τ_λ)
-            added_layer.t⁺⁺[:,:,iλ] = Diagonal(temp[iλ,:]);
-            added_layer.t⁻⁻[:,:,iλ] = Diagonal(temp[iλ,:]);
-        end
-    end
-
-    # @assert !any(isnan.(added_layer.t⁺⁺))
-    
-    # If this TOA, just copy the added layer into the bottom composite layer
-    if (iz == 1)
-        for ims = 1:length(sensor_levels)
-            if(ims==0)
-                composite_layer.botT⁺⁺[ims][:], composite_layer.botT⁻⁻[ims][:] = 
-                    Array(added_layer.t⁺⁺), Array(added_layer.t⁻⁻)
-                composite_layer.botR⁻⁺[ims][:], composite_layer.botR⁺⁻[ims][:] = 
-                    Array(added_layer.r⁻⁺), Array(added_layer.r⁺⁻)
-                composite_layer.botJ₀⁺[ims][:], composite_layer.botJ₀⁻[ims][:] = 
-                    Array(added_layer.j₀⁺), Array(added_layer.j₀⁻)
-            else
-                composite_layer.topT⁺⁺[ims][:], composite_layer.topT⁻⁻[ims][:] = 
-                    Array(added_layer.t⁺⁺), Array(added_layer.t⁻⁻)
-                composite_layer.topR⁻⁺[ims][:], composite_layer.topR⁺⁻[ims][:] = 
-                    Array(added_layer.r⁻⁺), Array(added_layer.r⁺⁻)
-                composite_layer.topJ₀⁺[ims][:], composite_layer.topJ₀⁻[ims][:] = 
[…]
