+++ src/CoreRT/rt_helper_functions.jl
@@ -14,16 +14,16 @@ function get_scattering_interface(scattering_interface,scatter, iz)
         scattering_interface = scatter ? ScatteringInterface_11() : ScatteringInterface_00()
     
     # Not the first layer (not TOA)
-    elseif !(scattering_interface isa ScatteringInterface_00)
-
+    else
+        
         # If kn was 1, then toggle between 0/0 and 0/1 
         # Else, toggle between 1/0 and 1/1
         scattering_interface = (scattering_interface isa ScatteringInterface_00) ? 
                                     (!scatter ? ScatteringInterface_00() : ScatteringInterface_01()) : 
                                     (!scatter ? ScatteringInterface_10() : ScatteringInterface_11())
     end
-
-    return scattering_interface
+    @show scattering_interface
+    return scattering_interface #ScatteringInterface_11() # 
 end
 
 "Minimum number of doublings needed to reach an optical depth τ_end, starting with an optical depth dτ.
