### 2022-10-16 f53351c
Fix bracket location in elemental
