### 2022-10-16 90b59aa
Remove output in HDRF comp
