### 2022-10-16 9e5ec35
Fix output save in RAMI
