### 2022-10-16 67c11e5
Simplify mu
