### 2022-10-16 f9841dd
Precompute Z matrices
