+++ src/CoreRT/rt_run_canopy.jl
@@ -84,12 +84,15 @@ function rt_run_canopy(RS_type::AbstractRamanType,
     
         #TODO: if RS_type!=noRS, create ϖ_λ₁λ₀, i_λ₁λ₀, fscattRayl, Z⁺⁺_λ₁λ₀, Z⁻⁺_λ₁λ₀ (for input), and ieJ₀⁺, ieJ₀⁻, ieR⁺⁻, ieR⁻⁺, ieT⁻⁻, ieT⁺⁺, ier⁺⁻, ier⁻⁺, iet⁻⁻, iet⁺⁺ (for output)
     #getRamanSSProp(RS_type, λ, grid_in)
-    
+    println("Prepping Canopy")
+    #@show BiLambMod,  Array(qp_μN), LAD
+    Zup, Zdown = CanopyOptics.precompute_Zazi(BiLambMod, Array(qp_μN), LAD)
+    #@show maximum(Zup)
     println("Finished initializing arrays")
 
     # Loop over fourier moments
     for m = 0:max_m - 1
-
+        
         println("Fourier Moment: ", m, "/", max_m-1)
         # Azimuthal weighting
         weight = m == 0 ? FT(0.5) : FT(1.0)
@@ -100,20 +103,39 @@ function rt_run_canopy(RS_type::AbstractRamanType,
             constructCoreOpticalProperties(RS_type,iBand,m,model);
         
         
-        #@show BiLambMod
-        𝐙⁺⁺, 𝐙⁻⁺ = CanopyOptics.compute_Z_matrices_aniso(BiLambMod, Array(qp_μN), LAD, m)    
+        𝐙⁺⁺, 𝐙⁻⁺ = CanopyOptics.compute_Z_matrices_aniso(BiLambMod, Array(qp_μN), LAD, Zup, Zdown, m)    
+        #𝐙⁺⁺, 𝐙⁻⁺ = CanopyOptics.compute_Z_matrices(BiLambMod, Array(qp_μN), LAD,  m)    
+        if m<4
+            @show Array(wt_μN') * 𝐙⁺⁺ + Array(wt_μN') * 𝐙⁻⁺
+        end  
+        
         # This basically multiplies with G again, needs to be fixed later (or removed from compute_Z_matrices)
         G1 = arr_type(CanopyOptics.G(Array(qp_μN), LAD))
+        #G1 .= 0.5# G1 *2
         #@show G1
-        canopyCore = CoreRT.CoreDirectionalScatteringOpticalProperties(arr_type(LAI*ones(FT, nSpec)), arr_type(ϖ_canopy*ones(FT,nSpec)), arr_type(𝐙⁺⁺), arr_type(𝐙⁻⁺), G1)
-        #@show canopyCore.ϖ
+        @time canopyCore = CoreRT.CoreDirectionalScatteringOpticalProperties(arr_type(LAI*ones(FT, nSpec)), arr_type(ϖ_canopy*ones(FT,nSpec)), arr_type(𝐙⁺⁺), arr_type(𝐙⁻⁺), G1)
+        #canopyCore = CoreRT.CoreScatteringOpticalProperties(arr_type(LAI*ones(FT, nSpec)), arr_type(ϖ_canopy*ones(FT,nSpec)), arr_type(𝐙⁺⁺), arr_type(𝐙⁻⁺))
+        #@show canopyCore.ϖ, canopyCore.τ, typeof(canopyCore )
         # Add Canopy at the bottom here:
+        #layer_opt_props =  [layer_opt_props; canopyCore]
+        a = layer_opt_props[end]
+        #a.τ .= 1.5
+        #if m>-1
+        #canopyCore.Z⁺⁺ .= a.Z⁺⁺[:,:,1]
[…]
