### 2022-10-19 2fc2ae2
Improve elemental_inelastic
