+++ src/CoreRT/CoreKernel/elemental_inelastic.jl
@@ -140,15 +140,14 @@ end
                 if abs(dτ_λ[n₀]-dτ_λ[n₁])>1.e-6
                     iet⁺⁺[i,j,n₁,Δn] = 
                         ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
-                        (exp(-dτ_λ[n₀] / qp_μN[i]) - exp(-dτ_λ[n₁] / qp_μN[i]))/
+                        (exp(-dτ_λ[n₁] / qp_μN[i]) - exp(-dτ_λ[n₀] / qp_μN[i]))/
                         (1 - (dτ_λ[n₁]/dτ_λ[n₀]))                         
                 else    
                     iet⁺⁺[i,j,n₁,Δn] = 
-                        ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
-                        (1 - exp(-dτ_λ[n₀] / qp_μN[j]))
+                        dτ_λ[n₀] * ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * 
+                        Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
+                        exp(-dτ_λ[n₀] / qp_μN[i])
                 end
-            else
-                iet⁺⁺[i,j,n₁,Δn] = 0.0
             end
         else
             #@show  qp_μN[i], qp_μN[j]  
@@ -156,7 +155,8 @@ end
             # (𝑖 ≠ 𝑗)
             iet⁺⁺[i,j,n₁,Δn] = 
                 ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,j] * 
-                (1 / ( (qp_μN[i]/qp_μN[j]) - (dτ_λ[n₁]/dτ_λ[n₀]) )) * wct2[j] * 
+                (1 / ( (qp_μN[i]/qp_μN[j]) - (dτ_λ[n₁]/dτ_λ[n₀]) )) * 
+                wct2[j] * 
                 (exp(-dτ_λ[n₁] / qp_μN[i]) - exp(-dτ_λ[n₀] / qp_μN[j]))
         end
     else
@@ -253,12 +253,13 @@ end
                 if abs(dτ_λ[n₀]-dτ_λ[n₁])>1.e-6
                     iet⁺⁺[i,j,n₁,1] = 
                         ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
-                        (exp(-dτ_λ[n₀] / qp_μN[i]) - exp(-dτ_λ[n₁] / qp_μN[i]))/
+                        (exp(-dτ_λ[n₁] / qp_μN[i]) - exp(-dτ_λ[n₀] / qp_μN[i]))/
                         (1 - (dτ_λ[n₁]/dτ_λ[n₀]))  
                 else    
                     iet⁺⁺[i,j,n₁,1] = 
-                        ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
-                        (1 - exp(-dτ_λ[n₀] / qp_μN[j]))   
+                        dτ_λ[n₀] * ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * 
+                        Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
[…]
