+++ src/CoreRT/CoreKernel/elemental_inelastic_plus.jl
@@ -318,12 +318,12 @@ end
                     if abs(dτ[n₀]-dτ[n₁])>1.e-6
                         iet⁺⁺[i,j,n₁,1] += 
                             ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
-                            (exp(-dτ[n₀] / qp_μN[i]) - exp(-dτ[n₁] / qp_μN[i]))/
+                            (exp(-dτ[n₁] / qp_μN[i]) - exp(-dτ[n₀] / qp_μN[i]))/
                             (1 - (dτ[n₁]/dτ[n₀]))  
                     else    
                         iet⁺⁺[i,j,n₁,1] += 
-                            ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
-                            (1 - exp(-dτ[n₀] / qp_μN[j]))   
+                            dτ[n₀] * ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,i] * wct2[i] *
+                            exp(-dτ[n₀] / qp_μN[j])   
                     end
                 else
                     iet⁺⁺[i,j,n₁,1] += 0.0
@@ -334,7 +334,8 @@ end
                 # (𝑖 ≠ 𝑗)
                 iet⁺⁺[i,j,n₁,1] += 
                         ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_λ₁λ₀[i,j] * 
-                        (1 / ( (qp_μN[i]/qp_μN[j]) - (dτ[n₁]/dτ[n₀]) )) * wct2[j] * 
+                        (1 / ( (qp_μN[i]/qp_μN[j]) - (dτ[n₁]/dτ[n₀]) )) * 
+                        wct2[j] * 
                         (exp(-dτ[n₁] / qp_μN[i]) - exp(-dτ[n₀] / qp_μN[j]))
             end
         else
@@ -456,13 +457,14 @@ end
             # J₀⁺ = 0.25*(1+δ(m,0)) * ϖ(λ) * Z⁺⁺ * I₀ * (dτ(λ)/μ₀) * exp(-dτ(λ)/μ₀)
             if abs(dτ[n₀]-dτ[n₁])>1.e-6
                 ieJ₀⁺[i, 1, n₁, 1] = 
-                        (exp(-dτ[n₀] / qp_μN[i]) - exp(-dτ[n₁] / qp_μN[i])) /
-                        ((dτ[n₁]/dτ[n₀])-1) * 
-                        ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_I₀ * wct02
+                        ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_I₀ * wct02 *
+                        (exp(-dτ[n₁] / qp_μN[i]) - exp(-dτ[n₀] / qp_μN[i])) /
+                        ((1-dτ[n₁]/dτ[n₀])) 
+                        
             else
                 ieJ₀⁺[i, 1, n₁, 1] = 
-                        wct02 * ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_I₀ * 
-                        (1 - exp(-dτ[n₀] / qp_μN[i_start]))
+                        dτ[n₀] *  wct02 * ϖ_λ₁λ₀[Δn] * fscattRayl[n₀] * Z⁺⁺_I₀ * 
[…]
