### 2022-10-19 50f92a7
Fix parts of canopy G integration, EB not perfect
