### 2022-10-19 ba09d3c
Fix parts of canopy G integration, EB not perfect
