### 2022-10-24 #4575 @pdebuyl
Thanks @RupeshJey  for the update!

I have Julia 1.5 and tried to install according to the instructions: (`]` key then `add vSmartMOM`) and have this error:

```
(@v1.5) pkg> add vSmartMOM
  Resolving package versions...
ERROR: Unsatisfiable requirements detected for package DelimitedFiles [8bb1440f]:
 DelimitedFiles [8bb1440f] log:
 ├─possible versions are: 1.5.4 or uninstalled
 └─found to have no compatible versions left with vSmartMOM [7ba11eeb]
   └─vSmartMOM [7ba11eeb] log:
     ├─possible versions are: 0.5.0 or uninstalled
     └─restricted to versions * by an explicit requirement, leaving only versions 0.5.0
```

Can you remove the data tables from the article file? It is better to have a link to the documentation, not only because tables in pdfs are hard to use but also to keep with JOSS' guidelines to have brief articles.

The citations yurchenko2012exomol:2012 roth-63 man2010hitemp:2010 have a syntax issue.
