### 2022-10-24 7fde3d6
Add G to tau sum
