### 2022-10-24 060aca2
Add plots to Canopy test
