+++ test/sandbox/oco_canopy_fullAtmAlbedo.jl
@@ -70,7 +70,7 @@ for Alb = 0.0:0.1:0.6
         #iBand = 3
         lAlb = [0.384,0.211,0.061]
         # Canopy Stuff!
-        LD = CanopyOptics.spherical_leaves()
+        LD = CanopyOptics.uniform_leaves()
         #LD = CanopyOptics.planophile_leaves2()
         leaf = LeafProspectProProperties{Float64}(Cw=0.006);
         LAI = 5.0
@@ -115,33 +115,87 @@ for iBand in eachindex(results_canopy)
     push!(wl, oco_sounding.SpectralGrid[oco_sounding.BandID[iBand2]])
     @show iBand
 end
+albedos = 0.0:0.1:0.6
+for i in eachindex(albedos)
+    iAlb = i
+    iBand = 1 + (iAlb-1)*3
+    #plot(wl[iBand], results_canopy_conv[iBand], label="Canopy")
+    #plot!(wl[iBand], results_NoCanopySameP_conv[iBand],label="No Canopy")
+    r1 = results_NoCanopySameP_conv[iBand]./results_NoCanopyLowP_conv[iBand]
+    r2 = results_canopy_conv[iBand]./results_NoCanopyLowP_conv[iBand]
+    plot(wl[iBand]*1e3, r1./maximum(r1), label="NoC_highP/noC_lowP", lw=1.5, alpha=0.5)
+    plot!(wl[iBand]*1e3, r2./maximum(r2), label="Integrated Canopy / No Canopy", lw=1.5, alpha=0.5)
+    xlabel!("Wavelength (nm)")
+    savefig("/home/cfranken/canopy/pdrdf_O2band"*string(albedos[i]*10) * ".pdf")
+    savefig("/home/cfranken/canopy/pdrdf_O2band"*string(albedos[i]*10) * ".png")
 
-iAlb = 4
+    iBand = 2+ (iAlb-1)*3
+    r1 = results_NoCanopySameP_conv[iBand]./results_NoCanopyLowP_conv[iBand]
+    r2 = results_canopy_conv[iBand]./results_NoCanopyLowP_conv[iBand]
+    plot(wl[iBand]*1e3, r1./maximum(r1), label="NoC_sameP/noC_lowP", lw=1.5, alpha=0.5)
+    plot!(wl[iBand]*1e3, r2./maximum(r2), label="Integrated Canopy / No Canopy", lw=1.5, alpha=0.5)
+    xlabel!("Wavelength (nm)")
+    savefig("/home/cfranken/canopy/pdrdf_WCO2band"*string(albedos[i]*10) * ".pdf")
+    savefig("/home/cfranken/canopy/pdrdf_WCO2band"*string(albedos[i]*10) * ".png")
+
+    iBand = 3+ (iAlb-1)*3
+    r1 = results_NoCanopySameP_conv[iBand]./results_NoCanopyLowP_conv[iBand]
+    r2 = results_canopy_conv[iBand]./results_NoCanopyLowP_conv[iBand]
+    plot(wl[iBand]*1e3, r1./maximum(r1), label="NoC_sameP/noC_lowP", lw=1.5, alpha=0.5)
+    plot!(wl[iBand]*1e3, r2./maximum(r2), label="Integrated Canopy / No Canopy", lw=1.5, alpha=0.5)
+    xlabel!("Wavelength (nm)")
+    savefig("/home/cfranken/canopy/pdrdf_SCO2band."*string(albedos[i]*10) * ".pdf")
[…]
