+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.5.0"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 CUDAKernels = "72cfdca4-0801-4ab0-bf6a-d52aa10adc57"
 CanopyOptics = "a18e34a6-5dbe-4f38-a44b-e5141852e7a7"
+ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
 DataInterpolations = "82cc6244-b520-54b8-b5a6-8a565e85f1d0"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 DiffResults = "163ba53b-c6d8-5494-b064-1a9d43ac40c5"

[…]
