### 2022-10-24 2ea7fd4
Use CuArray in MultiSensor (reduce allocations)
