### 2022-10-24 8b18b09
Use CuArray in MultiSensor (reduce allocations)
