### 2022-10-26 978a39c
Added CUDA Canopy speedup
