### 2022-10-26 497187b
Added CUDA Canopy speedup
