### 2022-10-26 2779144
Added CanopyOptics compat
