### 2022-10-27 241b04b
update extractEffectiveProps calls
