### 2022-10-27 3cc0ec4
update extractEffectiveProps calls
