### 2022-10-27 33fcdd6
Update architecture
