+++ src/CoreRT/model_from_parameters.jl
@@ -97,9 +97,9 @@ function model_from_parameters(params::vSmartMOM_Parameters)
     end
 
     # Rayleigh optical properties calculation
-    greek_rayleigh = Scattering.get_greek_rayleigh(params.depol)
-    @show params.depol
-    @show greek_rayleigh
+    greek_cabannes = Vector{vSmartMOM.Scattering.GreekCoefs{Float64}}()
+    greek_rayleigh = Vector{vSmartMOM.Scattering.GreekCoefs{Float64}}()
+    ϖ_Cabannes = zeros(n_bands)
     # Remove rayleight for testing:
     #τ_rayl = [zeros(params.float_type,length(params.spec_bands[i]), length(profile.T)) for i=1:n_bands];
     τ_rayl = [zeros(params.float_type,1,length(profile.T)) for i=1:n_bands];
@@ -119,10 +119,21 @@ function model_from_parameters(params::vSmartMOM_Parameters)
         #Suniti: the following two lines are temporary. this Cabannes depolarization applies only to the Earth's atmosphere
         # It has been added to make sure that the code sees the same Rayleigh cross section, regardless of elastic or inelastic RT 
         # Needs better (more general) formulation 
-        depol_Cabannes = 0.028
+        νₘ = 0.5*(params.spec_bands[i_band][1]+params.spec_bands[i_band][end])
+        λₘ = 1.e7/νₘ
+        @show i_band
+        ϖ_Cabannes[i_band], γ_air_Cabannes, γ_air_Rayleigh = 
+            InelasticScattering.compute_γ_air_Rayleigh!(λₘ)
+        @show ϖ_Cabannes[i_band]
+        depol_air_Cabannes = 2γ_air_Cabannes/(1+γ_air_Cabannes)
+        depol_air_Rayleigh = 2γ_air_Rayleigh/(1+γ_air_Rayleigh)
+        a = Scattering.get_greek_rayleigh(depol_air_Cabannes)
+        push!(greek_cabannes, a)
+        a = Scattering.get_greek_rayleigh(depol_air_Rayleigh)
+        push!(greek_rayleigh, a)
         τ_rayl[i_band]   .= getRayleighLayerOptProp(profile.p_half[end], 
-            (mean(curr_band_λ)), 
-            depol_Cabannes, profile.vcd_dry);
+            λₘ/1000., 
+            depol_air_Rayleigh, profile.vcd_dry);
 
         #=
         # Original form
@@ -228,10 +239,12 @@ function model_from_parameters(params::vSmartMOM_Parameters)
     end
 
     # Check the floating-type output matches specified FT
-
+@show size(ϖ_Cabannes), typeof(ϖ_Cabannes)
     # Return the model 
     return vSmartMOM_Model(params, 
-                        aerosol_optics,  
+                        aerosol_optics, 
+                        ϖ_Cabannes, 
+                        greek_cabannes,  
                         greek_rayleigh, 
                         quad_points, 
                         τ_abs, 
@@ -291,7 +304,11 @@ function model_from_parameters(RS_type::Union{VS_0to1_plus, VS_1to0_plus},
     #@show params.spec_bands
 
     # Rayleigh optical properties calculation
-    greek_rayleigh = Scattering.get_greek_rayleigh(params.depol)
+    #greek_rayleigh = Scattering.get_greek_rayleigh(params.depol)
+    # Rayleigh optical properties calculation
+    ϖ_Cabannes = zeros(n_bands)
[…]
