+++ src/Inelastic/raman_atmo_prop.jl
@@ -61,13 +61,13 @@ function getRamanSSProp!(RS_type::RRS, depol, λ, grid_in)
     atmo_σ_Rayl = compute_optical_Rayl(λ, n2, o2)
     RS_type.greek_raman = get_greek_raman(RS_type, n2, o2)
     #get_greek_raman!(RS_type, n2, o2)
-    RS_type.ϖ_Cabannes .= compute_ϖ_Cabannes(RS_type, depol, λ, n2, o2)
+    RS_type.ϖ_Cabannes .= compute_ϖ_Cabannes(RS_type, λ)
     # @show RS_type.ϖ_Cabannes
     # determine RRS cross-sections to λ₀ from nSpecRaman wavelengths around λ₀  
     index_raman_grid, atmo_σ_RRS = compute_optical_RS!(RS_type, grid_in, λ, n2, o2)
     # declare ϖ_Raman to be a grid of length raman grid
     #RS_type.ϖ_λ₁λ₀ = atmo_σ_RRS[end:-1:1]/atmo_σ_Rayl * (1-RS_type.ϖ_Cabannes[1])/sum(atmo_σ_RRS[end:-1:1]/atmo_σ_Rayl) #the grid gets inverted because the central wavelength is now seen as the recipient of RRS from neighboring source wavelengths
-    RS_type.ϖ_λ₁λ₀ = (atmo_σ_RRS[end:-1:1]/sum(atmo_σ_RRS[end:-1:1])) * (1-RS_type.ϖ_Cabannes[1]) #the grid gets inverted because the central wavelength is now seen as the recipient of RRS from neighboring source wavelengths
+    RS_type.ϖ_λ₁λ₀ = (atmo_σ_RRS[end:-1:1]/atmo_σ_Rayl) #the grid gets inverted because the central wavelength is now seen as the recipient of RRS from neighboring source wavelengths
     @show RS_type.ϖ_λ₁λ₀
     @show sum(RS_type.ϖ_λ₁λ₀)
     @show RS_type.ϖ_Cabannes
@@ -93,7 +93,7 @@ function getRamanSSProp!(RS_type::RRS_plus, depol)
     fscattRayl = zeros(FT, nBands)
     n_Raman = 0
     atmo_σ_Rayl = compute_optical_Rayl(λ, n2, o2)
-    ϖ_Cabannes[iB] = compute_ϖ_Cabannes(RS_type, depol, λ, n2, o2)
+    ϖ_Cabannes[iB] = compute_ϖ_Cabannes(RS_type, λ)
     for iB = 1:nBands
         _grid_in = grid_in[iB]
         λ = nm_per_m/(0.5*(_grid_in[1]+_grid_in[end]))
@@ -183,7 +183,7 @@ function getRamanSSProp!(
         
         if iB==1
             @show InelasticScattering.compute_ϖ_Cabannes(RS_type, depol, λ_inc, n2, o2)
-            ϖ_Cabannes[iB] = InelasticScattering.compute_ϖ_Cabannes(RS_type, depol, λ_inc, n2, o2)
+            ϖ_Cabannes[iB] = InelasticScattering.compute_ϖ_Cabannes(RS_type, λ_inc)
             #@show ϖ_Cabannes
         else
             ϖ_Cabannes[iB] = 1.
@@ -324,7 +324,7 @@ _grid_in = grid_in[iB]
 λ = nm_per_m/(0.5*(_grid_in[1]+_grid_in[end]))
 
 if iB==1
-    ϖ_Cabannes[iB] = InelasticScattering.compute_ϖ_Cabannes(RS_type, depol, λ_inc, n2, o2)
[…]
