### 2022-10-28 #4575 @pdebuyl
@RupeshJey there is also an issue with the copyright text in the README. The "all rights reserved" and the provisions for commercial use are not aligned with the Apache license. Please only keep the mention of apache licensing. You may keep the acknowledgement of course.
