### 2022-10-31 RemoteSensingTools/vSmartMOM.jl#127 @arjunsavel
Thanks for making these changes! The paper looks excellent!
