### 2022-10-31 38f94c3
CompatHelper: bump compat for CUDAKernels to 0.4, (keep existing compat)
