+++ Project.toml
@@ -35,7 +35,7 @@ YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
 [compat]
 CUDA = "3"
-CUDAKernels = "0.2, 0.3"
+CUDAKernels = "0.2, 0.3, 0.4"
 DataInterpolations = "3.6"
 DiffResults = "1.0"
 Distributions = "0.23, 0.24, 0.25"
