### 2022-10-31 666ab0f
CompatHelper: bump compat for DocStringExtensions to 0.9, (keep existing compat)
