+++ Project.toml
@@ -39,7 +39,7 @@ CUDAKernels = "0.2, 0.3"
 DataInterpolations = "3.6"
 DiffResults = "1.0"
 Distributions = "0.23, 0.24, 0.25"
-DocStringExtensions = "0.8"
+DocStringExtensions = "0.8, 0.9"
 FastGaussQuadrature = "0.4"
 ForwardDiff = "0.10"
 InstrumentOperator = "0.1"
