+++ Project.toml
@@ -47,7 +47,7 @@ Interpolations = "0.12, 0.13"
 JLD2 = "0.1, 0.2, 0.3, 0.4"
 JSON = "0.21"
 KernelAbstractions = "0.6, 0.7"
-NCDatasets = "0.11"
+NCDatasets = "0.11, 0.12"
 NNlib = "0.8"
 NetCDF = "0.10, 0.11"
 NetCDF_jll = "400.701.400 - 400.702.400"
