### 2022-10-31 9a96308
CompatHelper: bump compat for Polynomials to 3, (keep existing compat)
