+++ Project.toml
@@ -52,7 +52,7 @@ NNlib = "0.8"
 NetCDF = "0.10, 0.11"
 NetCDF_jll = "400.701.400 - 400.702.400"
 Parameters = "0.12"
-Polynomials = "2"
+Polynomials = "2, 3"
 ProgressMeter = "1.3"
 SpecialFunctions = "2"
 StaticArrays = "1.2"
