### 2022-10-31 1a8c118
CompatHelper: bump compat for KernelAbstractions to 0.8, (keep existing compat)
