+++ Project.toml
@@ -46,7 +46,7 @@ InstrumentOperator = "0.1"
 Interpolations = "0.12, 0.13"
 JLD2 = "0.1, 0.2, 0.3, 0.4"
 JSON = "0.21"
-KernelAbstractions = "0.6, 0.7"
+KernelAbstractions = "0.6, 0.7, 0.8"
 NCDatasets = "0.11"
 NNlib = "0.8"
 NetCDF = "0.10, 0.11"
