+++ Project.toml
@@ -43,7 +43,7 @@ DocStringExtensions = "0.8"
 FastGaussQuadrature = "0.4"
 ForwardDiff = "0.10"
 InstrumentOperator = "0.1"
-Interpolations = "0.12, 0.13"
+Interpolations = "0.12, 0.13, 0.14"
 JLD2 = "0.1, 0.2, 0.3, 0.4"
 JSON = "0.21"
 KernelAbstractions = "0.6, 0.7"
