### 2022-11-01 e3ed9ab
Remove tables and fix exomol reference
