+++ paper.md
@@ -60,7 +60,7 @@ The package has a modular architecture, allowing users to import just the specif
 
 ![Sample absorption spectrum of CO2 with 0.01 cm^-1 step size resolution, calculated using Absorption.jl](joss_2.png)
 
-**Absorption.jl** enables absorption cross-section calculations of atmospheric gases at different pressures, temperatures, and wavelengths. It uses the HITRAN [@Gordon:2017] database for calculations. For exoplanets and brown dwarfs, developments are underway to obtain spectra from the ExoMOL/HITEMP databases [@yurchenko2012exomol:2012; @rothman2010hitemp:2010], as described in @Sanghavi:2019. While the module enables lineshape calculations from scratch, it also allows users to create and save an interpolator object at specified wavelength, pressure, and temperature grids. The module also supports auto-differentiation of the profile, with respect to pressure and temperature. Calculations can be computed either on CPU or GPU (CUDA).
+**Absorption.jl** enables absorption cross-section calculations of atmospheric gases at different pressures, temperatures, and wavelengths. It uses the HITRAN [@Gordon:2017] database for calculations. For exoplanets and brown dwarfs, developments are underway to obtain spectra from the ExoMOL [@yurchenko2012exomol:2012] and HITEMP [@rothman2010hitemp:2010] databases, as described in @Sanghavi:2019. While the module enables lineshape calculations from scratch, it also allows users to create and save an interpolator object at specified wavelength, pressure, and temperature grids. The module also supports auto-differentiation of the profile, with respect to pressure and temperature. Calculations can be computed either on CPU or GPU (CUDA).
 
 ![Sample scattering phase functions of aerosols (I $\rightarrow$ I and the I $\rightarrow$ Q transition), calculated using Scattering.jl ($r_m$ = 10.0 $\mu$m, $\sigma$ = 1.1, $n_r$ = 1.3, $n_i$ = 0.0, $\lambda$ = 0.65 $\mu$m). Positive values are <span style="color:green">*green*</span>, and negative values are <span style="color:red">*red*</span>. ](joss_3.png)
 
@@ -68,7 +68,7 @@ The package has a modular architecture, allowing users to import just the specif
 
 # Benchmarks
 
-Standard reference tables from the literature [@Natraj:2009] are used to validate **vSmartMOM.jl** simulation output. The data is available in **Appendix 1**, in **Tables 1-3**. The key result is that simulated reflectance output from vSmartMOM.jl closely matches published standard values – always within 0.1% for I, 0.0005 for Q, and 0.0001 for U. 
[…]
