### 2022-11-01 #4575 @RupeshJey
@arjunsavel @jimmielin Thank you both for your time and effort in reviewing this article, and for your service to the open-source community as well!

@pdebuyl I am investigating the DelimitedFiles issue. Separately, our compatibility requirements specify that vSmartMOM is only available for Julia 1.6+. I wonder why it allows you to try to install it in the first place, since you are on Julia 1.5...

On the paper side, I have removed the reference tables and will include them in our documentation instead. The exomol/hitemp references are also fixed.

@cfranken Christian, I remember we went through the export process with JPL and confirmed we can use the open-source Apache 2.0 license. Did they ask us to include these extra provisions mentioned by Pierre, or did we include them from our side? Can we just remove them?
