+++ src/CoreRT/CoreKernel/doubling_inelastic.jl
@@ -51,6 +51,7 @@ function doubling_helper!(RS_type::RRS,
 
         # T⁺⁺(λ)[I - R⁺⁻(λ)R⁻⁺(λ)]⁻¹, for doubling R⁺⁻,R⁻⁺ and T⁺⁺,T⁻⁻ is identical
         batch_inv!(gp_refl, I_static .- r⁻⁺ ⊠ r⁻⁺)
+        
         @views tt⁺⁺_gp_refl[:] = t⁺⁺ ⊠ gp_refl
         if SFI
             # J⁺₂₁(λ) = J⁺₁₀(λ).exp(-τ(λ)/μ₀)
@@ -71,9 +72,9 @@ function doubling_helper!(RS_type::RRS,
                 #@show  n, ieJ₀⁺[1:pol_type.n:length(J₀⁻[:,1,1]),1,963,Δn]
                 #@show  n, ieJ₁⁺[1:pol_type.n:length(J₁⁻[:,1,1]),1,963,Δn]
             end
-        #@show  n, J₁⁺[1:pol_type.n:length(J₁⁻[:,1,1]),1,963]
-        ##@show  n, sum(ieJ₀⁺[1:pol_type.n:length(J₀⁻[:,1,1]),1,963,:], dims=2)
-        #@show  n, sum(ieJ₁⁺[1:pol_type.n:length(J₁⁻[:,1,1]),1,963,:], dims=2)
+            #@show  n, J₁⁺[1:pol_type.n:length(J₁⁻[:,1,1]),1,963]
+            ##@show  n, sum(ieJ₀⁺[1:pol_type.n:length(J₀⁻[:,1,1]),1,963,:], dims=2)
+            #@show  n, sum(ieJ₁⁺[1:pol_type.n:length(J₁⁻[:,1,1]),1,963,:], dims=2)
             #@show size(ieJ₁⁺)
             @timeit "precomp" tmp1 = gp_refl ⊠  (J₀⁺ + r⁻⁺ ⊠ J₁⁻)
             @timeit "precomp" tmp2 = gp_refl ⊠  (J₁⁻ + r⁻⁺ ⊠ J₀⁺)
@@ -85,7 +86,9 @@ function doubling_helper!(RS_type::RRS,
                 #@timeit "mati" @views tmp5 = ier⁻⁺[:,:,n₁,Δn] ⊠ r⁻⁺[:,:,n₀]
                 #@show size(tmp3)
                 #@show length(n₁), length(n₀), length(n₁_), length(n₀_)
-                @inbounds @views ieJ₀⁺[:,:,n₁,Δn] = 
+
+            
+                @inbounds @views tmp3 = 
                                 ieJ₁⁺[:,:,n₁,Δn] + 
                                 (tt⁺⁺_gp_refl[:,:,n₁] ⊠ 
                                 (ieJ₀⁺[:,:,n₁,Δn] + 
@@ -95,7 +98,7 @@ function doubling_helper!(RS_type::RRS,
                                 ier⁻⁺[:,:,n₁,Δn] ⊠ r⁻⁺[:,:,n₀]) ⊠ 
                                 tmp1[:,:,n₀])) + 
                                 iet⁺⁺[:,:,n₁,Δn] ⊠ tmp1[:,:,n₀]
-                @inbounds @views ieJ₀⁻[:,:,n₁,Δn] = 
+                @inbounds @views tmp4 = 
                                 ieJ₀⁻[:,:,n₁,Δn] + 
                                 (tt⁺⁺_gp_refl[:,:,n₁] ⊠ 
[…]
