+++ test/rami/RamiGasI_12.yaml
@@ -16,11 +16,11 @@ radiative_transfer:
   # Polarization type (Stokes_I, Stokes_IQ, Stokes_IQU, Stokes_IQUV)
   polarization_type:  Stokes_I()
   # Maximum m (number of Fourier moments to loop over)
-  max_m:              5
+  max_m:              20
   # Exclusion angle for forward peak in fitting procedure (degrees)
   Δ_angle:            2.0
   # Trunction length for legendre terms
-  l_trunc:            50
+  l_trunc:            20
   # Depolarization factor
   depol:              0.029
   # Floating point type for calculations (Float32, Float64)
