+++ test/rami/RamiGas.yaml
@@ -17,11 +17,11 @@ radiative_transfer:
   # Polarization type (Stokes_I, Stokes_IQ, Stokes_IQU, Stokes_IQUV)
   polarization_type:  Stokes_IQU()
   # Maximum m (number of Fourier moments to loop over)
-  max_m:              3
+  max_m:              20
   # Exclusion angle for forward peak in fitting procedure (degrees)
   Δ_angle:            2.0
   # Trunction length for legendre terms
-  l_trunc:            10
+  l_trunc:            20
   # Depolarization factor
   depol:              0.03
   # Floating point type for calculations (Float32, Float64)
