### 2022-11-07 dc939c0
CompatHelper: bump compat for FastGaussQuadrature to 0.5, (keep existing compat)
