+++ Project.toml
@@ -40,7 +40,7 @@ DataInterpolations = "3.6"
 DiffResults = "1.0"
 Distributions = "0.23, 0.24, 0.25"
 DocStringExtensions = "0.8"
-FastGaussQuadrature = "0.4"
+FastGaussQuadrature = "0.4, 0.5"
 ForwardDiff = "0.10"
 InstrumentOperator = "0.1"
 Interpolations = "0.12, 0.13"
