### 2022-11-07 #4575 @RupeshJey
@pdebuyl The copyright has been [updated](https://github.com/RemoteSensingTools/vSmartMOM.jl#copyright-notice)!
