### 2022-11-10 d74296b
Enable Polarization for all RAMI cases
