### 2022-11-10 7f97864
Enable Polarization for all RAMI cases
