### 2022-11-10 9d584b8
Fixed polarized output (index for sun was off)
