### 2022-11-12 #4575 @pdebuyl
Going through the final steps.
1. Testing the examples myself, I have a trouble with the undefined `model` variable in the absorption tutorial and with the `model_*` placeholder that should be replaced by an actual variable.
2. There are citation syntax issues (ExoMOL and HITEMP). Please check all reference for syntax.
3. The license issue is resolved, thanks.
4. Using with julia 1.8 works fine. Is there a way to prevent installation with julia 1.6?
5. Can you confirm the version number for the published version for this article?
6. Can you archive the code to zenodo with the same metadata (title etc) as for the article and post the DOI here?
