+++ paper.bib
@@ -219,7 +219,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   publisher={Oxford University Press Oxford, UK}
 }
 
-@article{yurchenko2012exomol,
+@article{yurchenko2012exomol:2012,
   title={Exomol: molecular line lists for exoplanet and other atmospheres},
   author={Yurchenko, SN and Tennyson, J},
   journal={European Astronomical Society Publications Series},
@@ -240,7 +240,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   publisher={IOP Publishing}
 }
 
-@article{rothman2010hitemp,
+@article{rothman2010hitemp:2010,
   title={HITEMP, the high-temperature molecular spectroscopic database},
   author={Rothman, Laurence S and Gordon, IE and Barber, RJ and Dothe, H and Gamache, Robert R and Goldman, A and Perevalov, VI and Tashkun, SA and Tennyson, J},
   journal={Journal of Quantitative Spectroscopy and Radiative Transfer},
