+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"EditURL = \"https://github.com/RemoteSensingTools/vSmartMOM.jl/blob/main/docs/src/pages/tutorials/Tutorial_Scattering.jl\"","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/#Scattering:-Mie-Phase-Function-Tutorial","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"","category":"section"},{"location":"pages/tutorials/Tutorial_Scattering/#Introduction","page":"Scattering: Mie Phase Function Tutorial","title":"Introduction","text":"","category":"section"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"In the following tutorial, we will walk through how to compute phase function matrices from Mie theory.","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"Along the way, we will also calculate greek coefficients (which are essentially the legendre decomposition of the phase matrix components), which will be required for radiative transfer calculations.","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/#Performing-the-phase-function-calculations","page":"Scattering: Mie Phase Function Tutorial","title":"Performing the phase function calculations","text":"","category":"section"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"First let's use the required packages","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"using vSmartMOM.Scattering\nusing Distributions\nusing Plots","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"Now, we define the aerosol size distribution and properties. We only support univariate aerosols for now, but will add multivariate aerosols soon (we can use mixture models easily).","category":"page"},{"location":"pages/tutorials/Tutorial_Scattering/","page":"Scattering: Mie Phase Function Tutorial","title":"Scattering: Mie Phase Function Tutorial","text":"μ     = 0.3;             ## Log-normal median radius [μm]\nσ     = 2.0;             ## Log-normal stddev of radius\nnᵣ    = 1.3;             ## Real part of refractive index\nnᵢ    = 0.0;             ## Imag part of refractive index (sign changed, use only + here)\nr_max = 30.0;            ## Maximum radius [μm]\nnquad_radius  = 2500;    ## Number of qu
[…]
