+++ Project.toml
@@ -41,7 +41,7 @@ DiffResults = "1.0"
 Distributions = "0.23, 0.24, 0.25"
 DocStringExtensions = "0.8"
 FastGaussQuadrature = "0.4"
-ForwardDiff = "0.10"
+ForwardDiff = "0.10.25"
 InstrumentOperator = "0.1"
 Interpolations = "0.12, 0.13"
 JLD2 = "0.1, 0.2, 0.3, 0.4"
