### 2022-11-19 441e32f
Test Julia 1.8 as well
