+++ .github/workflows/AutomatedGPUTests.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: self-hosted
     strategy:
       matrix:
-        julia-version: ['1.6.0', '1.7.0', '1.8.0']
+        julia-version: ['1.6.0', '1.7.0']
     
     steps:
       - uses: actions/checkout@v2
