+++ .github/workflows/AutomatedTests.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        julia-version: ['1.6.0', '1.7.0', '1.8.0']
+        julia-version: ['1.6.0', '1.7.0']
         os: [ubuntu-latest, windows-latest, macos-latest]
     
     steps:
