### 2022-11-20 c099891
Resume Julia 1.8 tests
