+++ versions.js
@@ -1,8 +1,9 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.0",
   "v0.4",
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.4.3";
+var DOCUMENTER_NEWEST = "v1.0.0";
 var DOCUMENTER_STABLE = "stable";
