+++ Project.toml
@@ -1,7 +1,7 @@
 name = "vSmartMOM"
 uuid = "7ba11eeb-0a61-4a04-a413-bf612cc2007e"
 authors = ["Rupesh Jeyaram <rjeyaram@caltech.edu> and contributors"]
-version = "1.0.0"
+version = "1.0.1"
 
 [deps]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
