### 2022-11-20 #4575 @RupeshJey
Hi Pierre, thank you for confirming these last steps!
1. Ah I see how this can be confusing. I have updated the documentation to include a line to let users know to replace `model_*` in both instances to whichever model they would like to use from Step 2.
2. These citation issues are resolved, and I have gone through the paper to double check that all the citations are accurate and appear properly in the generated pdf.
3. Thank you!
4. I believe you had installation issues when using Julia 1.5, no? This is already the correct behavior, because our package only supports Julia 1.6+. I installed Julia 1.5 and tried to install the package, but was correctly not allowed to install vSmartMOM. Installation using 1.6+ works fine, and I have also set up unit tests on Julia 1.8, so there should be no issues across these latest versions (1.6+).
5. Yes! The latest package version is **v1.0.1**, and this has been updated on the Julia package, GitHub release, and zenodo link below.
6. Yep! I used the GitHub integration, so it didn't ask me to provide all the metadata – just the GitHub link and repository access. If they click the GitHub link in the reference, it takes them directly to the repository. I think this is the appropriate way for GitHub repositories, but please let me know if I'm wrong. The DOI is here: [DOI: 10.5281/zenodo.7339527](https://doi.org/10.5281/zenodo.7339527)
