### 2022-11-23 #4575 @pdebuyl
Hi @RupeshJey you need to edit the metadata. The title is now "RemoteSensingTools/vSmartMOM.jl: v1.0.1". It should be "vSmartMOM.jl: an Open-Source Julia Package for Atmospheric Radiative Transfer and Remote Sensing Tools".

There are also more authors to the zenodo archive. Please set non-authors of the paper as contributors.
