+++ v1.0.1-3/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v1.0.1-3";
