+++ v1.0.1-3/pages/tutorials/Tutorial_Scattering.jl
@@ -1,0 +1,134 @@
+# # Scattering: Mie Phase Function Tutorial
+# 
+# ### Introduction
+# In the following tutorial, we will walk through how to compute phase function matrices from 
+# Mie theory.
+
+# Along the way, we will also calculate greek coefficients (which are essentially the 
+# legendre decomposition of the phase matrix components), which will be required for radiative 
+# transfer calculations. 
+
+# ---
+
+# ### Performing the phase function calculations
+
+# First let's use the required packages
+
+using vSmartMOM.Scattering
+using Distributions
+using Plots
+#----------------------------------------------------------------------------
+
+# Now, we define the aerosol size distribution and properties. We only support univariate
+# aerosols for now, but will add multivariate aerosols soon (we can use mixture models easily). 
+
+μ     = 0.3;             ## Log-normal median radius [μm]
+σ     = 2.0;             ## Log-normal stddev of radius
+nᵣ    = 1.3;             ## Real part of refractive index
+nᵢ    = 0.0;             ## Imag part of refractive index (sign changed, use only + here)
+r_max = 30.0;            ## Maximum radius [μm]
+nquad_radius  = 2500;    ## Number of quadrature points for integrating of size dist.
+
+## Create a Size Distribution (using Julia's Distributions package)
+size_distribution = LogNormal(log(μ), log(σ));
+
+#----------------------------------------------------------------------------
+
+# Create the aerosol
+aero = Aerosol(size_distribution, nᵣ, nᵢ)
+#----------------------------------------------------------------------------
+
+# Now, we define some scattering and truncation properties: 
+
+λ = 0.55;                             ## Incident wavelength [μm]
+polarization_type = Stokes_IQUV();    ## Polarization type 
+
+## Trunction length for legendre terms
+l_max = 20;
+
+## Exclusion angle for forward peak (in fitting procedure) `[degrees]`
+Δ_angle = 2;
+
+## Truncation type 
+truncation_type   = δBGE(l_max, Δ_angle);      
+#----------------------------------------------------------------------------
+
+## Create a Mie model, using the Siewert method NAI2
+model_NAI2 = make_mie_model(NAI2(), aero, λ, polarization_type, truncation_type, r_max, nquad_radius);
+#----------------------------------------------------------------------------
+
+## Compute aerosol optical properties:
+aerosol_optics_NAI2 = compute_aerosol_optical_properties(model_NAI2);
+#----------------------------------------------------------------------------
+
+# ### Plotting the greek coefficients
+# (These are basically giving us the legendre decomposition of the phase matrix components)
+
+using Parameters
+@unpack α,β,γ,δ,ϵ,ζ = aerosol_optics_NAI2.greek_coefs;
+p1 = plot(α,  title="α");
+p2 = plot(β,  title="β");
+p3 = plot(γ,  title="γ");
+p4 = plot(δ,  title="δ");
+p5 = plot(ϵ,  title="ϵ");
+p6 = plot(ζ,  title="ζ");
+plot(p1, p2, p3, p4, p5, p6, layout=(3, 2), legend=false)
+xlims!(0,100)
[…]
