### 2022-11-29 7376422
Fix bugs for Rayleigh = Cabannes+RRS match
