### 2022-11-29 #4575 @cfranken
I checked the doi links and these are the correct DOIs. It seems it doesn't
accept letters or brackets but the DOI links are correct

On Tue, Nov 29, 2022 at 1:40 PM Pierre de Buyl ***@***.***>
wrote:

> Woops, @RupeshJey <https://github.com/RupeshJey> please check all
> references.
>
