### 2022-11-30 2d47970
Update paper.bib
Added and edited DOIs
