+++ paper.bib
@@ -5,6 +5,7 @@
       eprint={1607.07892},
       archivePrefix={arXiv},
       journal = {arXiv:1607.07892 [cs.MS]},
+      doi={10.48550/arXiv.1607.07892},
       primaryClass={cs.MS}
 }
 
@@ -41,7 +42,7 @@
     year = {2017},
     note = {HITRAN2016 Special Issue},
     issn = {0022-4073},
-    doi = {https://doi.org/10.1016/j.jqsrt.2017.06.038},
+    doi = {10.1016/j.jqsrt.2017.06.038},
     url = {https://www.sciencedirect.com/science/article/pii/S0022407317301073},
     author = {I.E. Gordon and L.S. Rothman and C. Hill and R.V. Kochanov and Y. Tan and P.F. Bernath and M. Birk and V. Boudon and A. Campargue and K.V. Chance and B.J. Drouin and J.-M. Flaud and R.R. Gamache and J.T. Hodges and D. Jacquemart and V.I. Perevalov and A. Perrin and K.P. Shine and M.-A.H. Smith and J. Tennyson and G.C. Toon and H. Tran and V.G. Tyuterev and A. Barbe and A.G. Császár and V.M. Devi and T. Furtenbacher and J.J. Harrison and J.-M. Hartmann and A. Jolly and T.J. Johnson and T. Karman and I. Kleiner and A.A. Kyuberis and J. Loos and O.M. Lyulin and S.T. Massie and S.N. Mikhailenko and N. Moazzen-Ahmadi and H.S.P. Müller and O.V. Naumenko and A.V. Nikitin and O.L. Polyansky and M. Rey and M. Rotger and S.W. Sharpe and K. Sung and E. Starikova and S.A. Tashkun and J. Vander Auwera and G. Wagner and J. Wilzewski and P. Wcisło and S. Yu and E.J. Zak},
     keywords = {HITRAN, Spectroscopic database, Molecular spectroscopy, Molecular absorption, Spectroscopic line parameters, Absorption cross-sections, Collision-induced absorption, Aerosols},
@@ -123,7 +124,7 @@ volume = {251},
 pages = {112053},
 year = {2020},
 issn = {0034-4257},
-doi = {https://doi.org/10.1016/j.rse.2020.112053},
+doi = {10.1016/j.rse.2020.112053},
 url = {https://www.sciencedirect.com/science/article/pii/S0034425720304235},
 author = {Suniti Sanghavi and Robert Nelson and Christian Frankenberg and Michael Gunson},
 keywords = {OCO-2, XCO, Aerosol, O A-band, Weak and strong CO bands, Polarimetry, Multi-angle viewing, Climatology},
@@ -134,7 +135,6 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
    title={Photopolarimetric Characteristics of Brown Dwarfs. I. Uniform Cloud Decks},
    volume={866},
    ISSN={1538-4357},
-   url={http://dx.doi.org/10.3847/1538-4357/aadf94},
    DOI={10.3847/1538-4357/aadf94},
    number={1},
    journal={The Astrophysical Journal},
@@ -184,6 +184,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   number={1513},
   pages={183--197},
   year={1969},
+  doi={10.1098/rspa.1969.0187}
   publisher={The Royal Society London}
 }
 
@@ -195,6 +196,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   number={1513},
   pages={199--216},
   year={1969},
+  doi={10.1098/rspa.1969.0188},
[…]
