### 2022-11-30 #4575 @cfranken
ok, I got it now. I updated the bib file and managed to find DOIs for most
of the papers. Thanks for checking.

On Tue, Nov 29, 2022 at 3:48 PM Pierre de Buyl ***@***.***>
wrote:

> DOIs in the bib file must be only "10.xxx" and not the doi link "
> https://doi.org/10.xxx". Even though the latter form is accepted as the
> official value for the DOI registry, it should not be used in the bib file.
> The links in the pdf have an extra "https" in front and the doi checker
> does not accept them.
>
> Also, some papers don't have a DOI whereas there is likely one. Please
> check all.
>
