+++ paper.bib
@@ -184,7 +184,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   number={1513},
   pages={183--197},
   year={1969},
-  doi={10.1098/rspa.1969.0187}
+  doi={10.1098/rspa.1969.0187},
   publisher={The Royal Society London}
 }
 
@@ -207,7 +207,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   volume={26},
   number={5},
   pages={963--972},
-  doi={10.1175/1520-0469(1969)026%3C0963:DSTORT%3E2.0.CO;2}
+  doi={10.1175/1520-0469(1969)026%3C0963:DSTORT%3E2.0.CO;2},
   year={1969}
 }
 
@@ -281,7 +281,7 @@ abstract = {We discuss the accuracy of the correlated-k method for radiation cal
   volume={55},
   number={29},
   pages={8236--8247},
-  doi={10.1364/AO.55.008236},
+  doi={10.1364/AO.55.008236}
 }
 
 @article{spurr:2011,
