+++ paper.bib
@@ -207,7 +207,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   volume={26},
   number={5},
   pages={963--972},
-  doi={10.1175/1520-0469(1969)026%3C0963:DSTORT%3E2.0.CO;2},
+  doi={10.1175/1520-0469(1969)026<0963:DSTORT>2.0.CO;2},
   year={1969}
 }
 
@@ -232,6 +232,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
   volume={58},
   pages={243--248},
   year={2012},
+  doi={10.1051/eas/1258039},
   publisher={EDP Sciences}
 }
 
