### 2022-11-30 #4575 @cfranken
On Wed, Nov 30, 2022 at 12:52 AM Pierre de Buyl ***@***.***>
wrote:

> The DOI for Hunt & Grant is 10.1175/1520-0469(1969)026<0963:DSTORT>2.0.CO
> ;2 some characters were lost in the bib entry.
> The DOI for Yurchenko and Tennyson is 10.1051/eas/1258039
>
