### 2022-11-30 #4575 @pdebuyl
Hi @cfranken there are still a few inconsistencies in the bibliography. The delta in the delta-fit reference does not show properly and the journal name is repeated. Natraj 2009 lacks the article number. The JQSRT papers by Sanghavi lack the article number-page as well. For the HITRAN2016 paper, please list all authors or shorten with et al.
