+++ paper.bib
@@ -13,7 +13,7 @@
     author = {Sanghavi, Suniti and Davis, Anthony and Eldering, Annmarie},
     year = {2013},
     month = {09},
-    pages = {},
+      pages = {412-433},
     title = {vSmartMOM: A vector matrix operator method-based radiative transfer model linearized with respect to aerosol properties},
     volume = {133},
     journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
@@ -56,6 +56,7 @@ month = {12},
 pages = {},
 title = {Revisiting the Fourier expansion of Mie scattering matrices in generalized spherical functions},
 volume = {136},
+pages = {16-27},
 journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
 doi = {10.1016/j.jqsrt.2013.12.015}
 }
@@ -67,6 +68,8 @@ month = {02},
 pages = {},
 title = {Rayleigh Scattering in Planetary Atmospheres: Corrected Tables Through Accurate Computation of X and Y Functions},
 volume = {691},
+number = {2},
+pages = {1909},
 journal = {The Astrophysical Journal},
 doi = {10.1088/0004-637X/691/2/1909}
 }
@@ -75,12 +78,13 @@ doi = {10.1088/0004-637X/691/2/1909}
 author = {Sanghavi, Suniti and Stephens, Graeme},
 year = {2015},
 month = {03},
-pages = {},
-title = {Adaptation of the delta-m and δ-fit truncation methods to vector radiative transfer: Effect of truncation on radiative transfer accuracy},
+pages = {53-68},
+title = {Adaptation of the delta-m and $\delta$-fit truncation methods to vector radiative transfer: Effect of truncation on radiative transfer accuracy},
 volume = {159},
 journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
 doi = {10.1016/j.jqsrt.2015.03.007}
 }
+
 
 @article{Siewert:1982,
 author = {Siewert, C.},
@@ -249,16 +253,21 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
 }
 
 @article{rothman2010hitemp:2010,
-  title={HITEMP, the high-temperature molecular spectroscopic database},
-  author={Rothman, Laurence S and Gordon, IE and Barber, RJ and Dothe, H and Gamache, Robert R and Goldman, A and Perevalov, VI and Tashkun, SA and Tennyson, J},
-  journal={Journal of Quantitative Spectroscopy and Radiative Transfer},
-  volume={111},
-  number={15},
-  pages={2139--2150},
-  year={2010},
-  doi={10.1016/j.jqsrt.2010.05.001},
-  publisher={Elsevier}
+  title = {HITEMP, the high-temperature molecular spectroscopic database},
+journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+volume = {111},
+number = {15},
+pages = {2139-2150},
+year = {2010},
+note = {XVIth Symposium on High Resolution Molecular Spectroscopy (HighRus-2009)},
+issn = {0022-4073},
+doi = {10.1016/j.jqsrt.2010.05.001},
+url = {https://www.sciencedirect.com/science/article/pii/S002240731000169X},
+author = {L.S. Rothman and I.E. Gordon and R.J. Barber and H. Dothe and R.R. Gamache and A. Goldman and V.I. Perevalov and S.A. Tashkun and J. Tennyson},
+keywords = {Spectroscopic database, Molecular spectroscopy, Molecular absorption, Line parameters, High-temperature spectroscopy, HITEMP},
[…]
