### 2022-11-30 #4575 @cfranken
The HITRAN authors were all listed in the BIB file, we can't control the style for the pdf generation I suppose.
