+++ paper.bib
@@ -151,7 +151,7 @@ abstract = {We have analyzed the effect of aerosols on the retrieval of the dry 
 @ARTICLE{Sanghavi:2019,
        author = {{Sanghavi}, Suniti and {West}, Robert},
         title = "{Spectropolarimetric Characteristics of Brown Dwarfs. II. Uniform Clouds}",
-      journal = {\apj},
+      journal={The Astrophysical Journal},
      keywords = {brown dwarfs, opacity, planets and satellites: gaseous planets, polarization, radiative transfer, scattering},
          year = 2019,
         month = jun,
