### 2022-12-01 #4575 @pdebuyl
Sorry for the HITRAN authors, I had a similar issue recently where the bib file was truncated and I assumed this is what happened.
