### 2022-12-05 #4575 @kthyng
Hi @RupeshJey! I'm helping to wrap up this publication, however, I see some capitalization errors in your references. For example it looks like "Mie" should be capitalized but it isn't always. Also in your Grant et al references, there is an "i" instead of "I". These are just two examples, please look through your references in detail to correct all errors. You can preserve capitalization with {} around characters.
