### 2022-12-05 ae1758f
Update paper.bib
Keep capitalization in paper titles
