+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Revels:2016,
-      title={Forward-Mode Automatic Differentiation in Julia}, 
+      title={{Forward-Mode Automatic Differentiation in Julia}}, 
       author={Jarrett Revels and Miles Lubin and Theodore Papamarkou},
       year={2016},
       eprint={1607.07892},
@@ -14,7 +14,7 @@
     year = {2013},
     month = {09},
       pages = {412-433},
-    title = {vSmartMOM: A vector matrix operator method-based radiative transfer model linearized with respect to aerosol properties},
+    title = {{vSmartMOM: A vector matrix operator method-based radiative transfer model linearized with respect to aerosol properties}},
     volume = {133},
     journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
     doi = {10.1016/j.jqsrt.2013.09.004}
@@ -22,7 +22,7 @@
 
 @ARTICLE{1975ZeMet..25..357D,
        author = {{Domke}, H.},
-        title = "{Fourier expansion of the phase matrix for Mie scattering}",
+        title = "{{Fourier expansion of the phase matrix for Mie scattering}}",
       journal = {Zeitschrift fur Meteorologie},
      keywords = {Fourier Analysis, Light Scattering, Mie Scattering, Polarized Light, S Matrix Theory, Computer Techniques, Fourier Series, Legendre Functions, Scattering Coefficients, Scattering Cross Sections, Vectors (Mathematics), Physics (General)},
          year = 1975,
@@ -35,7 +35,7 @@
 }
 
 @article{Gordon:2017,
-    title = {The HITRAN2016 molecular spectroscopic database},
+    title = {{The HITRAN2016 molecular spectroscopic database}},
     journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
     volume = {203},
     pages = {3-69},
@@ -54,7 +54,7 @@ author = {Sanghavi, Suniti},
 year = {2013},
 month = {12},
 pages = {},
-title = {Revisiting the Fourier expansion of Mie scattering matrices in generalized spherical functions},
+title = {{Revisiting the Fourier expansion of Mie scattering matrices in generalized spherical functions}},
 volume = {136},
 pages = {16-27},
 journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
@@ -66,7 +66,7 @@ author = {Natraj, Vijay and Li, King-Fai and Yung, Yuk},
 year = {2009},
 month = {02},
 pages = {},
-title = {Rayleigh Scattering in Planetary Atmospheres: Corrected Tables Through Accurate Computation of X and Y Functions},
+title = {{Rayleigh Scattering in Planetary Atmospheres: Corrected Tables Through Accurate Computation of X and Y Functions}},
 volume = {691},
 number = {2},
 pages = {1909},
@@ -79,7 +79,7 @@ author = {Sanghavi, Suniti and Stephens, Graeme},
 year = {2015},
 month = {03},
 pages = {53-68},
-title = {Adaptation of the delta-m and $\delta$-fit truncation methods to vector radiative transfer: Effect of truncation on radiative transfer accuracy},
+title = {{Adaptation of the delta-m and $\delta$-fit truncation methods to vector radiative transfer: Effect of truncation on radiative transfer accuracy}},
 volume = {159},
 journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
 doi = {10.1016/j.jqsrt.2015.03.007}
@@ -91,14 +91,14 @@ author = {Siewert, C.},
 year = {1982},
 month = {04},
 pages = {195},
-title = {On the Phase Matrix Basic to the Scattering of Polarized Light},
+title = {{On the Phase Matrix Basic to the Scattering of Polarized Light}},
[…]
