### 2022-12-05 #4575 @cfranken
we added {} for the titles (bibtex should automatically keep capitalization for titles ideally, this has always been annoying!).
Thanks for checking, this is much more thorough than in other journals
