### 2022-12-06 #4575 @cfranken
I second @RupeshJey, this has been a great process at JOSS, other journals should learn from it (instead of charging an arm and a leg).
